refactor(media): split MediaService into focused sub-services
\`media.service.ts\` had grown to ~3100 lines and bundled every concern
of the media domain (import, metadata, query, mutation, rescan) into a
single class. Splits it into five focused services under a new
\`media-service/\` directory plus a small TMDB-mapping util, all wired
through \`MediaModule\`:

- MediaImportService — TMDB / TVDB import flows
- MediaMetadataService — metadata fetch + persistence (cast, crew, etc.)
- MediaQueryService — read-side queries (findAll, findOne, counts…)
- MediaMutationService — write-side updates (bulk, profiles, monitored)
- MediaRescanService — disk rescan + enrichImportedFile

\`MediaService\` stays in place as a thin facade so external callers keep
the same import path and method names — no controller or downstream-
module changes required.
