+++ backend/src/modules/media/media.service.ts
@@ -1,22 +1,9 @@
-import {
-  Injectable,
-  Logger,
-  NotFoundException,
-  ConflictException,
-  BadRequestException,
-  forwardRef,
-  Inject,
-} from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, DataSource, SelectQueryBuilder, Brackets } from 'typeorm';
+import { Repository } from 'typeorm';
 import { Media } from './entities/media.entity';
-import { MediaFile } from './entities/media-file.entity';
 import { Season } from './entities/season.entity';
 import { Episode } from './entities/episode.entity';
-import { DownloadHistory } from './entities/download-history.entity';
-import { MediaMetadata } from './entities/media-metadata.entity';
-import { Person } from './entities/person.entity';
 import { MediaCast } from './entities/media-cast.entity';
 import { MediaCrew } from './entities/media-crew.entity';
 import { CreateMediaDto } from './dto/create-media.dto';
@@ -28,2249 +15,188 @@ import { ImportMediaDto } from './dto/import-media.dto';
 import { UpdateMediaProfilesDto } from './dto/update-media-profiles.dto';
 import { BulkUpdateMediaDto } from './dto/bulk-update-media.dto';
 import { CalendarQueryDto } from './dto/calendar-query.dto';
-import { TmdbProvider } from '../metadata-providers/providers/tmdb.provider';
-import { MetadataProviderRegistry } from '../metadata-providers/metadata-provider.registry';
-import {
-  IMetadataProvider,
-  MetadataDetails,
-  SeasonDetails,
-} from '../metadata-providers/interfaces/metadata-provider.interface';
-import { MediaType, MediaStatus } from '../../common/enums';
-import { RequestLifecycleService } from '../requests/request-lifecycle.service';
-import { ProfilesService } from '../profiles/profiles.service';
-import { QualityProfile } from '../profiles/entities/quality-profile.entity';
-import { LanguageProfile } from '../profiles/entities/language-profile.entity';
-import { User } from '../users/entities/user.entity';
-import { Library } from '../libraries/entities/library.entity';
-import { LibrariesService } from '../libraries/libraries.service';
-import { NamingService } from '../scheduler/naming.service';
-import {
-  getAppQualityById,
-  APP_QUALITIES,
-} from '../../common/constants/app-qualities';
-import { rankFromQualityString } from './release-rejection.helper';
-
-import { ImageService } from '../images/image.service';
-import { SubtitleStreamService } from '../streaming/subtitle-stream.service';
-import {
-  ThumbnailService,
-  buildSpriteLabel,
-} from '../streaming/thumbnail.service';
-import { EmbeddedSubtitleService } from '../subtitles/embedded-subtitle.service';
-import { clearMediaCache } from '../../common/utils/media-cache.util';
-import { mapWithConcurrency } from '../../common/utils/concurrency';
-import { MediaServersService } from '../media-servers/media-servers.service';
-import { FfprobeService } from '../subtitles/ffprobe.service';
-import { SubtitlesService } from '../subtitles/subtitles.service';
-import * as fs from 'fs';
-import * as path from 'path';
-import { relativePathUnderMediaRoot } from '../../common/utils/media-path.util';
+import { MediaType } from '../../common/enums';
+import { MediaImportService } from './media-service/media-import.service';
+import { MediaMetadataService } from './media-service/media-metadata.service';
+import { MediaQueryService } from './media-service/media-query.service';
+import { MediaMutationService } from './media-service/media-mutation.service';
+import { MediaRescanService } from './media-service/media-rescan.service';
 
 /**
- * Subquery returning every media.id that the user (`:userId` bind) has
- * finished. Used by both `excludeWatched` (NOT IN …) and `onlyWatched`
- * (IN …) in `findAll`.
- *
- * - Movies: any completed playback for that media counts as watched.
- * - Series: watched only when every episode with a file (specials, season 0,
- *   excluded) has a completed playback row for this user. A series with no
- *   downloaded episodes never qualifies as watched.
+ * Façade in front of the media sub-services. External callers
+ * (controllers, schedulers, other modules) only see this surface, so
+ * the public method signatures here are load-bearing — changing them
+ * means changing every importer.
  */
-const WATCHED_MEDIA_IDS_SUBQUERY = `
-  SELECT DISTINCT ps."mediaId" AS id FROM playback_states ps
-  INNER JOIN media m ON m.id = ps."mediaId"
-  WHERE ps."userId" = :userId AND ps.completed = true AND m.type = 'movie'
-  UNION
-  SELECT m2.id FROM media m2
-  WHERE m2.type = 'series'
-  AND EXISTS (
-    SELECT 1 FROM seasons s
-    JOIN episodes e ON e."seasonId" = s.id
-    WHERE s."mediaId" = m2.id AND s."seasonNumber" > 0 AND e."hasFile" = true
-  )
-  AND NOT EXISTS (
-    SELECT 1 FROM seasons s
-    JOIN episodes e ON e."seasonId" = s.id
-    WHERE s."mediaId" = m2.id AND s."seasonNumber" > 0 AND e."hasFile" = true
-    AND NOT EXISTS (
-      SELECT 1 FROM playback_states ps2
-      WHERE ps2."userId" = :userId AND ps2."episodeId" = e.id AND ps2.completed = true
-    )
-  )
-`;
-
 @Injectable()
 export class MediaService {
-  private readonly log = new Logger(MediaService.name);
-
   constructor(
     @InjectRepository(Media)
     private readonly mediaRepo: Repository<Media>,
-    @InjectRepository(Season)
-    private readonly seasonRepo: Repository<Season>,
-    @InjectRepository(Episode)
-    private readonly episodeRepo: Repository<Episode>,
-    @InjectRepository(MediaFile)
-    private readonly mediaFileRepo: Repository<MediaFile>,
-    @InjectRepository(DownloadHistory)
-    private readonly historyRepo: Repository<DownloadHistory>,
-    @InjectRepository(Library)
-    private readonly libraryRepo: Repository<Library>,
-    private readonly libraries: LibrariesService,
-    @InjectRepository(MediaMetadata)
-    private readonly metadataRepo: Repository<MediaMetadata>,
-    @InjectRepository(Person)
-    private readonly personRepo: Repository<Person>,
-    @InjectRepository(MediaCast)
-    private readonly castRepo: Repository<MediaCast>,
-    @InjectRepository(MediaCrew)
-    private readonly crewRepo: Repository<MediaCrew>,
-    private readonly dataSource: DataSource,
-    private readonly tmdb: TmdbProvider,
-    private readonly providerRegistry: MetadataProviderRegistry,
-    private readonly config: ConfigService,
-    private readonly profiles: ProfilesService,
-    private readonly naming: NamingService,
-    private readonly embeddedSubtitle: EmbeddedSubtitleService,
-    private readonly mediaServers: MediaServersService,
-    private readonly ffprobe: FfprobeService,
-    private readonly subtitles: SubtitlesService,
-    private readonly imageService: ImageService,
-    private readonly subtitleStream: SubtitleStreamService,
-    private readonly thumbnailService: ThumbnailService,
-    @Inject(forwardRef(() => RequestLifecycleService))
-    private readonly requestLifecycle: RequestLifecycleService,
+    private readonly imports: MediaImportService,
+    private readonly metadata: MediaMetadataService,
+    private readonly query: MediaQueryService,
+    private readonly mutation: MediaMutationService,
+    private readonly rescan: MediaRescanService,
   ) {}
 
-  async importFromTmdb(
-    dto: ImportTmdbDto,
-    addedByUserId: number | null = null,
-  ): Promise<Media> {
-    const key = this.config.get<string>('TMDB_API_KEY', '');
-    if (!key?.trim()) {
-      throw new BadRequestException('TMDB API key is not configured');
-    }
+  // -- Imports ----------------------------------------------------------------
 
-    const existing = await this.mediaRepo.findOne({
-      where: { tmdbId: dto.tmdbId, type: dto.type },
-    });
-    if (existing) {
-      throw new ConflictException('This title is already in the library');
-    }
-
-    const qualityProfileId =
-      await this.profiles.resolveQualityProfileIdForImport(
-        dto.qualityProfileId,
-      );
-
-    const languageProfileId =
-      await this.profiles.resolveLanguageProfileIdForImport(
-        dto.languageProfileId,
-      );
-
-    const { libraryId } = await this.resolveImportTarget(dto.type, {
-      libraryId: dto.libraryId,
-    });
-
-    // Load folder format settings
-    const fmtKeys = [
-      'naming_movie_folder_format',
-      'naming_series_folder_format',
-    ];
-    const fmtRows: { key: string; value: string }[] =
-      await this.dataSource.query(
-        `SELECT key, value FROM app_settings WHERE key = ANY($1)`,
-        [fmtKeys],
-      );
-    const fmtMap = Object.fromEntries(fmtRows.map((r) => [r.key, r.value]));
-
-    if (dto.type === MediaType.MOVIE) {
-      const details = await this.tmdb.getMovieDetails(String(dto.tmdbId));
-      const movieFolderFormat =
-        fmtMap['naming_movie_folder_format'] ??
-        '{Original Title} ({Release Year})';
-      const folderName = this.naming.applyMovieFolderFormat(movieFolderFormat, {
-        title: details.title,
-        originalTitle: details.originalTitle,
-        year: details.year,
-        tmdbId: details.tmdbId,
-      });
-      return this.persistImportedMovie(
-        details,
-        qualityProfileId,
-        languageProfileId,
-        folderName,
-        libraryId,
-        addedByUserId,
-      );
-    }
-
-    const details = await this.tmdb.getTvShowDetails(String(dto.tmdbId));
-    const seasons = await this.tmdb.getTvShowSeasons(String(dto.tmdbId));
-    const seriesFolderFormat =
-      fmtMap['naming_series_folder_format'] ?? '{Series Title}';
-    const folderName = this.naming.applySeriesFolderFormat(seriesFolderFormat, {
-      seriesTitle: details.title,
-      originalTitle: details.originalTitle,
-      year: details.year,
-      tmdbId: details.tmdbId,
-    });
-    return this.persistImportedSeries(
-      details,
-      seasons,
-      qualityProfileId,
-      languageProfileId,
-      folderName,
-      libraryId,
-      addedByUserId,
-    );
+  importFromTmdb(dto: ImportTmdbDto, addedByUserId: number | null = null) {
+    return this.imports.importFromTmdb(dto, addedByUserId);
   }
 
-  /**
-   * Import media from any provider (TMDB, TVDB).
-   * Cross-references IDs between providers when possible.
-   */
-  async importMedia(
-    dto: ImportMediaDto,
-    addedByUserId: number | null = null,
-  ): Promise<Media> {
-    const provider = this.providerRegistry.resolve(dto.provider ?? null);
-
-    // Check for existing media (by any known ID)
-    const existingCheck: any[] = [];
-    if (dto.provider === 'tmdb' || !dto.provider) {
-      existingCheck.push({
-        tmdbId: parseInt(dto.externalId, 10),
-        type: dto.type,
-      });
-    }
-    if (dto.provider === 'tvdb') {
-      existingCheck.push({
-        tvdbId: parseInt(dto.externalId, 10),
-        type: dto.type,
-      });
-    }
-    if (existingCheck.length) {
-      const existing = await this.mediaRepo.findOne({ where: existingCheck });
-      if (existing) {
-        throw new ConflictException('This title is already in the library');
-      }
-    }
-
-    const qualityProfileId =
-      await this.profiles.resolveQualityProfileIdForImport(
-        dto.qualityProfileId,
-      );
-    const languageProfileId =
-      await this.profiles.resolveLanguageProfileIdForImport(
-        dto.languageProfileId,
-      );
-
-    const { libraryId } = await this.resolveImportTarget(dto.type, {
-      libraryId: dto.libraryId,
-    });
-
-    // Load folder format settings
-    const fmtKeys = [
-      'naming_movie_folder_format',
-      'naming_series_folder_format',
-    ];
-    const fmtRows: { key: string; value: string }[] =
-      await this.dataSource.query(
-        `SELECT key, value FROM app_settings WHERE key = ANY($1)`,
-        [fmtKeys],
-      );
-    const fmtMap = Object.fromEntries(fmtRows.map((r) => [r.key, r.value]));
-
-    // Fetch details from provider
-    if (dto.type === MediaType.MOVIE) {
-      const details = await provider.getMovieDetails(dto.externalId);
-      // Cross-reference: if we got a tvdbId but no tmdbId, try to resolve via TMDB
-      if (!details.tmdbId && details.tvdbId && this.tmdb.findByExternalId) {
-        const cross = await this.tmdb.findByExternalId(
-          'tvdb',
-          String(details.tvdbId),
-        );
-        if (cross) details.tmdbId = parseInt(cross.id, 10);
-      }
-      const movieFolderFormat =
-        fmtMap['naming_movie_folder_format'] ??
-        '{Original Title} ({Release Year})';
-      const folderName = this.naming.applyMovieFolderFormat(movieFolderFormat, {
-        title: details.title,
-        originalTitle: details.originalTitle,
-        year: details.year,
-        tmdbId: details.tmdbId,
-      });
-      return this.persistImportedMovie(
-        details,
-        qualityProfileId,
-        languageProfileId,
-        folderName,
-        libraryId,
-        addedByUserId,
-      );
-    }
-
-    const details = await provider.getTvShowDetails(dto.externalId);
-    const seasons = await provider.getTvShowSeasons(dto.externalId);
-    // Cross-reference
-    if (!details.tmdbId && details.tvdbId && this.tmdb.findByExternalId) {
-      const cross = await this.tmdb.findByExternalId(
-        'tvdb',
-        String(details.tvdbId),
-      );
-      if (cross) details.tmdbId = parseInt(cross.id, 10);
-    }
-    const seriesFolderFormat =
-      fmtMap['naming_series_folder_format'] ?? '{Series Title}';
-    const folderName = this.naming.applySeriesFolderFormat(seriesFolderFormat, {
-      seriesTitle: details.title,
-      originalTitle: details.originalTitle,
-      year: details.year,
-      tmdbId: details.tmdbId,
-    });
-    return this.persistImportedSeries(
-      details,
-      seasons,
-      qualityProfileId,
-      languageProfileId,
-      folderName,
-      libraryId,
-      addedByUserId,
-    );
+  importMedia(dto: ImportMediaDto, addedByUserId: number | null = null) {
+    return this.imports.importMedia(dto, addedByUserId);
   }
 
-  /**
-   * Resolve the destination library for an import.
-   *  - Explicit `libraryId` from DTO wins (validated against media type).
-   *  - Otherwise we fall back to the default library for the media type
-   *    (`isDefaultForMovies` / `isDefaultForSeries` flag).
-   *
-   * The library is also required to have a configured `path` — every
-   * import flow needs to know where to drop files.
-   */
-  private async resolveImportTarget(
-    type: MediaType,
-    dto: { libraryId?: number },
-  ): Promise<{ libraryId: number }> {
-    let library: Library | null = null;
-
-    if (dto.libraryId) {
-      library = await this.libraryRepo.findOne({
-        where: { id: dto.libraryId },
-      });
-      if (!library) {
-        throw new BadRequestException(`Library #${dto.libraryId} not found`);
-      }
-    }
-
-    if (!library) {
-      library = await this.libraries.getDefaultForType(type);
-    }
-
-    if (!library) {
-      throw new BadRequestException(
-        'No compatible library found. Set a default library for this media type in settings.',
-      );
-    }
-    if (!library.mediaTypes?.includes(type)) {
-      throw new BadRequestException(
-        `Library "${library.name}" does not accept ${type}`,
-      );
-    }
-    if (!library.path) {
-      throw new BadRequestException(
-        `Library "${library.name}" has no root path configured`,
-      );
-    }
-
-    return { libraryId: library.id };
+  create(dto: CreateMediaDto) {
+    return this.imports.create(dto);
   }
 
-  async create(dto: CreateMediaDto): Promise<Media> {
-    const media = this.mediaRepo.create(dto);
+  // -- Reads ------------------------------------------------------------------
 
-    const saved = await this.mediaRepo.save(media);
-    await this.updateSearchVector(saved.id);
-    return this.findOne(saved.id);
+  getCounts(accessibleLibraryIds: number[]) {
+    return this.query.getCounts(accessibleLibraryIds);
   }
 
-  async getCounts(
-    accessibleLibraryIds: number[],
-  ): Promise<{ movies: number; series: number }> {
-    const buildQb = (type: MediaType) => {
-      const qb = this.mediaRepo
-        .createQueryBuilder('media')
-        .where('media.type = :type', { type });
-      this.applyLibraryAcl(qb, accessibleLibraryIds);
-      return qb;
-    };
-    const [movies, series] = await Promise.all([
-      buildQb(MediaType.MOVIE).getCount(),
-      buildQb(MediaType.SERIES).getCount(),
-    ]);
-    return { movies, series };
+  getCountsByLibrary(accessibleLibraryIds: number[]) {
+    return this.query.getCountsByLibrary(accessibleLibraryIds);
   }
 
-  async getCountsByLibrary(
-    accessibleLibraryIds: number[],
-  ): Promise<Record<number, number>> {
-    const qb = this.mediaRepo
-      .createQueryBuilder('m')
-      .select('m."libraryId"', 'libraryId')
-      .addSelect('COUNT(*)::int', 'count')
-      .where('m."libraryId" IS NOT NULL')
-      .groupBy('m."libraryId"');
-    this.applyLibraryAcl(qb, accessibleLibraryIds);
-    const rows: { libraryId: number; count: number }[] = await qb.getRawMany();
-    return Object.fromEntries(rows.map((r) => [r.libraryId, r.count]));
+  getGenres(accessibleLibraryIds: number[]) {
+    return this.query.getGenres(accessibleLibraryIds);
   }
 
-  /**
-   * Aggregate distinct genres across the accessible libraries, with the
-   * total item count + up to 4 sample posters per genre (used by the
-   * library Genres tab to render a mosaic when posters are available).
-   * Skips media with null / empty `posterUrl` for the sample collection
-   * — they wouldn't render anything useful in the mosaic.
-   */
-  async getGenres(
-    accessibleLibraryIds: number[],
-  ): Promise<{ genre: string; count: number; posters: string[] }[]> {
-    if (accessibleLibraryIds.length === 0) return [];
-    // jsonb_array_elements_text unrolls the `media.genres` array so we
-    // can GROUP BY genre. The lateral join lets us aggregate poster
-    // URLs in the same pass; LIMIT 4 inside the array_agg keeps payload
-    // small without a per-group subquery.
-    const rows: { genre: string; count: number; posters: string[] }[] =
-      await this.mediaRepo.query(
-        `
-        SELECT g.genre,
-               COUNT(*)::int AS count,
-               COALESCE(
-                 (ARRAY_AGG(m."posterUrl" ORDER BY m.id)
-                  FILTER (WHERE m."posterUrl" IS NOT NULL))[1:4],
-                 ARRAY[]::text[]
-               ) AS posters
-        FROM media m
-        CROSS JOIN LATERAL jsonb_array_elements_text(m.genres) AS g(genre)
-        WHERE m."libraryId" = ANY($1)
-          AND m.genres IS NOT NULL
-          AND m.genres::text != '[]'
-        GROUP BY g.genre
-        ORDER BY g.genre ASC
-        `,
-        [accessibleLibraryIds],
-      );
-    return rows;
+  getCollections(accessibleLibraryIds: number[]) {
+    return this.query.getCollections(accessibleLibraryIds);
   }
 
-  async getCollections(
-    accessibleLibraryIds: number[],
-  ): Promise<{ id: number; name: string; count: number; posters: string[] }[]> {
-    if (accessibleLibraryIds.length === 0) return [];
-    const rows: {
-      id: number;
-      name: string;
-      count: number;
-      posters: string[];
-    }[] = await this.mediaRepo.query(
-      `
-        SELECT m."tmdbCollectionId"   AS id,
-               m."tmdbCollectionName" AS name,
-               COUNT(*)::int          AS count,
-               COALESCE(
-                 (ARRAY_AGG(m."posterUrl" ORDER BY m.id)
-                  FILTER (WHERE m."posterUrl" IS NOT NULL))[1:4],
-                 ARRAY[]::text[]
-               ) AS posters
-        FROM media m
-        WHERE m."libraryId" = ANY($1)
-          AND m."tmdbCollectionId" IS NOT NULL
-          AND m."tmdbCollectionName" IS NOT NULL
-        GROUP BY m."tmdbCollectionId", m."tmdbCollectionName"
-        HAVING COUNT(*) >= 2
-        ORDER BY m."tmdbCollectionName" ASC
-        `,
-      [accessibleLibraryIds],
-    );
-    return rows;
-  }
-
-  async findAll(
+  findAll(
     query: SearchMediaDto,
     userId?: number,
     accessibleLibraryIds: number[] = [],
-  ): Promise<{ data: Media[]; total: number }> {
-    const page = query.page ?? 1;
-    const limit = query.limit ?? 25;
-    const offset = (page - 1) * limit;
-
-    const qb = this.mediaRepo
-      .createQueryBuilder('media')
-      .leftJoinAndSelect('media.library', 'library')
-      .leftJoinAndSelect('media.qualityProfile', 'qualityProfile')
-      .leftJoinAndSelect('media.languageProfile', 'languageProfile')
-      .leftJoinAndSelect('media.files', 'files');
-
-    this.applyLibraryAcl(qb, accessibleLibraryIds);
-    this.applyFilters(qb, query);
-
-    if ((query.excludeWatched || query.onlyWatched) && userId) {
-      const op = query.onlyWatched ? 'IN' : 'NOT IN';
-      qb.andWhere(`media.id ${op} (${WATCHED_MEDIA_IDS_SUBQUERY})`, { userId });
-    }
-
-    if (query.requestedByMe && userId) {
-      // Approved requests set `media.addedById` to the requester, and direct
-      // imports set it to the importer — so a single column lookup covers
-      // both flows.
-      qb.andWhere('media."addedById" = :reqUserId', { reqUserId: userId });
-    }
-
-    if (query.q) {
-      this.applyFullTextSearch(qb, query.q);
-    }
-
-    // Whitelist + column mapping for ORDER BY. The raw `sortBy` value
-    // arrives from the client, so we must reject anything outside the
-    // allowed keys (otherwise the string ends up concatenated into the
-    // SQL — classic ORDER BY injection vector) and translate UI-side
-    // aliases like `added` to the actual physical column (`createdAt`,
-    // inherited from BaseEntity).
-    const SORT_BY_MAP: Record<string, string> = {
-      title: 'media.title',
-      year: 'media.year',
-      added: 'media.createdAt',
-      createdAt: 'media.createdAt',
-      rating: 'media.rating',
-    };
-    const sortBy = SORT_BY_MAP[query.sortBy ?? 'title'] ?? 'media.title';
-    const sortOrder = query.sortOrder === 'DESC' ? 'DESC' : 'ASC';
-    qb.orderBy(sortBy, sortOrder);
-
-    if (limit > 0) {
-      qb.skip(offset).take(limit);
-    }
-
-    const [data, total] = await qb.getManyAndCount();
-
-    // For series: attach episode stats
-    const seriesIds = data
-      .filter((m) => m.type === MediaType.SERIES)
-      .map((m) => m.id);
-    let episodeStatsMap = new Map<
-      number,
-      { totalEpisodes: number; downloadedEpisodes: number }
-    >();
-    if (seriesIds.length) {
-      const stats: { mediaId: number; total: string; downloaded: string }[] =
-        await this.dataSource.query(
-          `SELECT s."mediaId",
-                  COUNT(e.id) AS total,
-                  COUNT(mf.id) AS downloaded
-           FROM seasons s
-           JOIN episodes e ON e."seasonId" = s.id
-           LEFT JOIN media_files mf ON mf."episodeId" = e.id
-           WHERE s."mediaId" = ANY($1) AND s."seasonNumber" > 0
-           GROUP BY s."mediaId"`,
-          [seriesIds],
-        );
-      episodeStatsMap = new Map(
-        stats.map((s) => [
-          s.mediaId,
-          {
-            totalEpisodes: parseInt(s.total, 10),
-            downloadedEpisodes: parseInt(s.downloaded, 10),
-          },
-        ]),
-      );
-    }
-
-    let enriched = data.map((m) => {
-      const stats = episodeStatsMap.get(m.id);
-      return Object.assign(m, {
-        sizeOnDisk: (m.files ?? []).reduce((sum, f) => sum + Number(f.size), 0),
-        episodeStats: stats ?? undefined,
-      });
-    });
-
-    if (query.cutoffUnmet === true) {
-      // Cutoff comparison must reach below the cutoff *on the unit that gets
-      // downloaded* — the whole movie for movies, each individual episode
-      // for series. The previous implementation looked at all
-      // `media.files[]` collectively for series, so a single upgraded
-      // episode hid every other below-cutoff one. We also fold "totally
-      // missing" media (no file → rank 0) into this bucket and lean on
-      // `parseReleaseQuality` (via {@link rankFromQualityString}) so legacy
-      // stored quality strings are parsed instead of needing an exact
-      // `APP_QUALITIES.name` match.
-      const cutoffByMedia = new Map<number, number>();
-      for (const m of enriched) {
-        if (!m.qualityProfile) continue;
-        const cq = getAppQualityById(m.qualityProfile.cutoff);
-        if (cq) cutoffByMedia.set(m.id, cq.rank);
-      }
-      const seriesIds = enriched
-        .filter((m) => m.type === MediaType.SERIES && cutoffByMedia.has(m.id))
-        .map((m) => m.id);
-      const epBestByMedia = new Map<number, Map<number, number>>();
-      if (seriesIds.length) {
-        const epRows: {
-          mediaId: number;
-          epId: number;
-          quality: string | null;
-        }[] = await this.dataSource.query(
-          `SELECT s."mediaId" AS "mediaId",
-                  e.id      AS "epId",
-                  f."quality" AS "quality"
-             FROM seasons s
-             JOIN episodes e ON e."seasonId" = s.id
-             LEFT JOIN media_files f ON f."episodeId" = e.id
-            WHERE s."mediaId" = ANY($1)
-              AND s."seasonNumber" > 0
-              AND e.monitored = true`,
-          [seriesIds],
-        );
-        for (const row of epRows) {
-          let byEp = epBestByMedia.get(row.mediaId);
-          if (!byEp) {
-            byEp = new Map();
-            epBestByMedia.set(row.mediaId, byEp);
-          }
-          const rank = rankFromQualityString(row.quality);
-          const prev = byEp.get(row.epId);
-          if (prev == null || rank > prev) byEp.set(row.epId, rank);
-        }
-      }
-      enriched = enriched.filter((m) => {
-        const cutoffRank = cutoffByMedia.get(m.id);
-        if (cutoffRank == null) return false;
-        if (m.type === MediaType.MOVIE) {
-          let rank = 0;
-          for (const f of m.files ?? []) {
-            const r = rankFromQualityString(f.quality);
-            if (r > rank) rank = r;
-          }
-          return rank < cutoffRank;
-        }
-        const byEp = epBestByMedia.get(m.id);
-        if (!byEp || byEp.size === 0) return false;
-        for (const rank of byEp.values()) {
-          if (rank < cutoffRank) return true;
-        }
-        return false;
-      });
-    }
-
-    return { data: enriched, total };
+  ) {
+    return this.query.findAll(query, userId, accessibleLibraryIds);
   }
 
-  /** When `accessibleLibraryIds` is omitted the lookup is unscoped — used
-   *  by internal callers (admin approve path, schedulers) that have
-   *  already validated access. Controllers must pass the user's library
-   *  set so the result respects the ACL. */
-  async findByTmdbId(
+  findByTmdbId(
     tmdbId: number,
     type: MediaType,
     accessibleLibraryIds?: number[],
-  ): Promise<Media | null> {
-    const m = await this.mediaRepo.findOne({ where: { tmdbId, type } });
-    if (!m) return null;
-    if (accessibleLibraryIds) {
-      if (m.libraryId == null || !accessibleLibraryIds.includes(m.libraryId)) {
-        return null;
-      }
-    }
-    return m;
+  ) {
+    return this.query.findByTmdbId(tmdbId, type, accessibleLibraryIds);
   }
 
-  /**
-   * Throws NotFoundException when the media exists but is outside the user's
-   * accessible libraries — same shape as "not found" so we don't leak existence.
-   */
-  async assertAccessible(
-    mediaId: number,
-    accessibleLibraryIds: number[],
-  ): Promise<void> {
-    // `libraryId` is a @RelationId virtual — TypeORM's findOne + select
-    // can't project it. Use a raw query on the join column instead.
-    const row = await this.mediaRepo
-      .createQueryBuilder('m')
-      .select('m."libraryId"', 'libraryId')
-      .where('m.id = :mediaId', { mediaId })
-      .getRawOne<{ libraryId: number | null }>();
-    if (!row) throw new NotFoundException(`Media #${mediaId} not found`);
-    if (
-      row.libraryId == null ||
-      !accessibleLibraryIds.includes(row.libraryId)
-    ) {
-      throw new NotFoundException(`Media #${mediaId} not found`);
-    }
+  findOne(id: number) {
+    return this.query.findOne(id);
   }
 
-  /** Returns the parent media id for a season — used by ACL on season-scoped endpoints. */
-  async getMediaIdForSeason(seasonId: number): Promise<number> {
-    const s = await this.seasonRepo.findOne({ where: { id: seasonId } });
-    if (!s) throw new NotFoundException(`Season #${seasonId} not found`);
-    return s.mediaId;
+  getCast(mediaId: number): Promise<MediaCast[]> {
+    return this.query.getCast(mediaId);
   }
 
-  /** Returns the parent media id for an episode — used by ACL on episode-scoped endpoints. */
-  async getMediaIdForEpisode(episodeId: number): Promise<number> {
-    const e = await this.episodeRepo.findOne({
-      where: { id: episodeId },
-      relations: ['season'],
-    });
-    if (!e) throw new NotFoundException(`Episode #${episodeId} not found`);
-    return e.season.mediaId;
+  getCrew(mediaId: number): Promise<MediaCrew[]> {
+    return this.query.getCrew(mediaId);
   }
 
-  /** Adds `WHERE <alias>.libraryId IN (...)` to scope a query to the
-   *  caller's accessible libraries. Uses the query builder's own alias
-   *  so it works whether the caller aliased the media table as "media"
-   *  (default) or "m" (findAll / counts queries). An empty array short-
-   *  circuits to `1 = 0` so the caller still hits a real WHERE and
-   *  returns no rows. */
-  private applyLibraryAcl(
-    qb: SelectQueryBuilder<Media>,
-    accessibleLibraryIds: number[],
-  ): void {
-    if (accessibleLibraryIds.length === 0) {
-      qb.andWhere('1 = 0');
-      return;
-    }
-    const alias = qb.alias;
-    qb.andWhere(`${alias}."libraryId" IN (:...accessibleLibraryIds)`, {
-      accessibleLibraryIds,
-    });
+  getMediaIdForSeason(seasonId: number) {
+    return this.query.getMediaIdForSeason(seasonId);
   }
 
-  async getCast(mediaId: number): Promise<MediaCast[]> {
-    return this.castRepo.find({
-      where: { media: { id: mediaId } },
-      relations: ['person'],
-      order: { order: 'ASC' },
-    });
+  getMediaIdForEpisode(episodeId: number) {
+    return this.query.getMediaIdForEpisode(episodeId);
   }
 
-  async getCrew(mediaId: number): Promise<MediaCrew[]> {
-    return this.crewRepo.find({
-      where: { media: { id: mediaId } },
-      relations: ['person'],
-    });
+  assertAccessible(mediaId: number, accessibleLibraryIds: number[]) {
+    return this.query.assertAccessible(mediaId, accessibleLibraryIds);
   }
 
-  async findOne(id: number): Promise<Media> {
-    // relationLoadStrategy 'query' issues one SELECT per relation instead of
-    // a single LEFT JOIN — without it, a series with 200 episodes × 200 files
-    // × N tags hydrates the cartesian-product (40k+ rows) and the endpoint
-    // takes ~4 s. With separate queries each relation is bounded and the
-    // total drops to a few hundred ms.
-    const media = await this.mediaRepo.findOne({
-      where: { id },
-      relationLoadStrategy: 'query',
-      relations: [
-        'seasons',
-        'seasons.episodes',
-        'files',
-        'qualityProfile',
-        'languageProfile',
-        'library',
-        // Eager on the entity, but relationLoadStrategy 'query' above
-        // skips eager auto-load — list it explicitly so the detail
-        // endpoint exposes the extended TMDB fields (originalLanguage,
-        // productionCountries, productionCompanies, tagline, …).
-        'metadata',
-      ],
-    });
-    if (!media) {
-      throw new NotFoundException(`Media #${id} not found`);
-    }
-    if (media.seasons?.length) {
-      media.seasons.sort((a, b) => a.seasonNumber - b.seasonNumber);
-      for (const s of media.seasons) {
-        s.episodes?.sort((a, b) => a.episodeNumber - b.episodeNumber);
-      }
-    }
-    if (media.type === MediaType.SERIES && media.seasons?.length) {
-      const epIdsWithTrackedFile = new Set(
-        (media.files ?? [])
-          .map((f) => f.episodeId)
-          .filter((id): id is number => id != null && id > 0),
-      );
-      for (const s of media.seasons) {
-        for (const e of s.episodes ?? []) {
-          if (epIdsWithTrackedFile.has(e.id)) {
-            e.hasFile = true;
-          }
-        }
-      }
-    }
-    return media;
-  }
-
-  async update(id: number, dto: UpdateMediaDto): Promise<Media> {
-    const media = await this.findOne(id);
-    const wasMonitored = media.monitored;
-    const { path: _path, ...rest } = dto;
-
-    Object.assign(media, rest);
-
-    const saved = await this.mediaRepo.save(media);
-    await this.updateSearchVector(saved.id);
-    await this.requestLifecycle.onMediaMonitorChange(saved, wasMonitored);
-    return this.findOne(saved.id);
-  }
-
-  /**
-   * Reassign media to a different library. Picks a root folder inside the
-   * target library (most free space) and updates both FKs atomically.
-   */
-  async updateLibrary(id: number, libraryId: number): Promise<Media> {
-    await this.findOne(id);
-    const library = await this.libraryRepo.findOne({
-      where: { id: libraryId },
-    });
-    if (!library) {
-      throw new NotFoundException(`Library #${libraryId} not found`);
-    }
-    if (!library.path) {
-      throw new BadRequestException(
-        `Library "${library.name}" has no root path configured`,
-      );
-    }
-    await this.mediaRepo.update(id, { library });
-    return this.findOne(id);
-  }
-
-  async updateProfiles(
-    id: number,
-    dto: UpdateMediaProfilesDto,
-  ): Promise<Media> {
-    await this.findOne(id);
-    const patch: {
-      qualityProfile?: QualityProfile | null;
-      languageProfile?: LanguageProfile | null;
-    } = {};
-    if (dto.qualityProfileId !== undefined) {
-      if (dto.qualityProfileId !== null) {
-        await this.profiles.findOneQualityProfile(dto.qualityProfileId);
-        patch.qualityProfile = { id: dto.qualityProfileId } as QualityProfile;
-      } else {
-        patch.qualityProfile = null;
-      }
-    }
-    if (dto.languageProfileId !== undefined) {
-      if (dto.languageProfileId !== null) {
-        await this.profiles.findOneLanguageProfile(dto.languageProfileId);
-        patch.languageProfile = {
-          id: dto.languageProfileId,
-        } as LanguageProfile;
-      } else {
-        patch.languageProfile = null;
-      }
-    }
-    if (Object.keys(patch).length === 0) {
-      throw new BadRequestException(
-        'Provide at least one of qualityProfileId or languageProfileId',
-      );
-    }
-    await this.mediaRepo.update(
-      { id },
-      patch as Parameters<Repository<Media>['update']>[1],
-    );
-    return this.findOne(id);
-  }
-
-  async bulkUpdate(dto: BulkUpdateMediaDto): Promise<{ updated: number }> {
-    const patch: Partial<Record<string, unknown>> = {};
-
-    if (dto.qualityProfileId !== undefined) {
-      patch.qualityProfile =
-        dto.qualityProfileId === null
-          ? null
-          : ({ id: dto.qualityProfileId } as QualityProfile);
-    }
-    if (dto.languageProfileId !== undefined) {
-      patch.languageProfile =
-        dto.languageProfileId === null
-          ? null
-          : ({ id: dto.languageProfileId } as LanguageProfile);
-    }
-    if (dto.monitored !== undefined) {
-      patch.monitored = dto.monitored;
-    }
-    if (dto.libraryId !== undefined) {
-      patch.libraryId = dto.libraryId;
-    }
-
-    if (Object.keys(patch).length === 0) {
-      throw new BadRequestException('Provide at least one field to update');
-    }
-
-    const result = await this.mediaRepo
-      .createQueryBuilder()
-      .update(Media)
-      .set(patch)
-      .whereInIds(dto.ids)
-      .execute();
-
-    return { updated: result.affected ?? 0 };
-  }
-
-  async remove(id: number): Promise<void> {
-    const media = await this.findOne(id);
-    const title = media.title;
-    const mediaPath = media.path;
-    await this.requestLifecycle.onMediaRemoved(media);
-    await this.mediaRepo.remove(media);
-    void this.mediaServers.dispatch('media.deleted', {
-      title,
-      path: mediaPath,
-    });
-  }
-
-  // ---------------------------------------------------------------------------
-  // Calendar
-  // ---------------------------------------------------------------------------
-
-  async getCalendar(
+  getCalendar(
     dto: CalendarQueryDto,
     accessibleLibraryIds: number[],
     userId?: number,
   ) {
-    // TypeORM may return PostgreSQL `date` columns as Date objects.
-    // Normalise to YYYY-MM-DD string to avoid timezone shifts.
-    function toDateStr(v: unknown): string | null {
-      if (!v) return null;
-      if (v instanceof Date) {
-        const y = v.getUTCFullYear();
-        const m = String(v.getUTCMonth() + 1).padStart(2, '0');
-        const d = String(v.getUTCDate()).padStart(2, '0');
-        return `${y}-${m}-${d}`;
-      }
-      if (typeof v === 'string') return v.slice(0, 10);
-      if (typeof v === 'number' || typeof v === 'bigint')
-        return String(v).slice(0, 10);
-      return null;
-    }
-    let start: string;
-    let end: string;
-    if (dto.start && dto.end) {
-      start = dto.start.slice(0, 10);
-      end = dto.end.slice(0, 10);
-    } else {
-      const now = new Date();
-      const y = now.getFullYear();
-      const m = now.getMonth();
-      start = `${y}-${String(m + 1).padStart(2, '0')}-01`;
-      const last = new Date(y, m + 1, 0).getDate();
-      end = `${y}-${String(m + 1).padStart(2, '0')}-${String(last).padStart(2, '0')}`;
-    }
-
-    type CalendarEntry = {
-      id: number;
-      mediaId: number;
-      title: string;
-      type: 'movie' | 'series';
-      event: string;
-      date: string;
-      posterUrl: string | null;
-      status: string;
-      year: number;
-      seasonNumber?: number;
-      episodeNumber?: number;
-      episodeTitle?: string;
-      hasFile?: boolean;
-    };
-
-    const results: CalendarEntry[] = [];
-
-    // 1. Movies — one entry per event type with a date in range
-    if (!dto.type || dto.type === MediaType.MOVIE) {
-      const moviesQb = this.mediaRepo
-        .createQueryBuilder('m')
-        .where('m.type = :type', { type: MediaType.MOVIE })
-        .andWhere(
-          new Brackets((qb) => {
-            qb.where('m.inCinemas BETWEEN :start AND :end', { start, end })
-              .orWhere('m.digitalRelease BETWEEN :start AND :end', {
-                start,
-                end,
-              })
-              .orWhere('m.physicalRelease BETWEEN :start AND :end', {
-                start,
-                end,
-              })
-              .orWhere('m.releaseDate BETWEEN :start AND :end', { start, end });
-          }),
-        );
-      if (dto.monitoredOnly) {
-        moviesQb.andWhere('m.monitored = true');
-      }
-      if (dto.requestedByMe && userId) {
-        moviesQb.andWhere('m."addedById" = :calUserId', { calUserId: userId });
-      }
-      if (accessibleLibraryIds.length === 0) {
-        moviesQb.andWhere('1 = 0');
-      } else {
-        moviesQb.andWhere('m.libraryId IN (:...accessibleLibraryIds)', {
-          accessibleLibraryIds,
-        });
-      }
-      const movies = await moviesQb.getMany();
-
-      const eventFields: { field: keyof Media; event: string }[] = [
-        { field: 'inCinemas', event: 'cinema' },
-        { field: 'digitalRelease', event: 'digital' },
-        { field: 'physicalRelease', event: 'physical' },
-      ];
-
-      for (const m of movies) {
-        let hasSpecificDate = false;
-        for (const { field, event } of eventFields) {
-          const d = toDateStr(m[field]);
-          if (d && d >= start && d <= end) {
-            hasSpecificDate = true;
-            results.push({
-              id: m.id,
-              mediaId: m.id,
-              title: m.title,
-              type: 'movie',
-              event,
-              date: d,
-              posterUrl: m.posterUrl,
-              status: m.status,
-              year: m.year,
-            });
-          }
-        }
-        // Fallback to generic releaseDate if no specific dates
-        const rd = toDateStr(m.releaseDate);
-        if (!hasSpecificDate && rd && rd >= start && rd <= end) {
-          results.push({
-            id: m.id,
-            mediaId: m.id,
-            title: m.title,
-            type: 'movie',
-            event: 'release',
-            date: rd,
-            posterUrl: m.posterUrl,
-            status: m.status,
-            year: m.year,
-          });
-        }
-      }
-    }
-
-    // 2. Episodes with airDate in range
-    if (!dto.type || dto.type === MediaType.SERIES) {
-      const epQb = this.episodeRepo
-        .createQueryBuilder('ep')
-        .innerJoinAndSelect('ep.season', 'season')
-        .innerJoinAndSelect('season.media', 'media')
-        .where('ep.airDate BETWEEN :start AND :end', { start, end })
-        .orderBy('ep.airDate', 'ASC');
-      if (dto.monitoredOnly) {
-        epQb.andWhere('media.monitored = true').andWhere('ep.monitored = true');
-      }
-      if (dto.requestedByMe && userId) {
-        epQb.andWhere('media."addedById" = :calUserId', { calUserId: userId });
-      }
-      if (accessibleLibraryIds.length === 0) {
-        epQb.andWhere('1 = 0');
-      } else {
-        epQb.andWhere('media.libraryId IN (:...accessibleLibraryIds)', {
-          accessibleLibraryIds,
-        });
-      }
-      const episodes = await epQb.getMany();
-
-      for (const ep of episodes) {
-        results.push({
-          id: ep.id,
-          mediaId: ep.season.media.id,
-          title: ep.season.media.title,
-          type: 'series',
-          event: 'airing',
-          date: toDateStr(ep.airDate) ?? ep.airDate,
-          posterUrl: ep.season.media.posterUrl,
-          status: ep.season.media.status,
-          year: ep.season.media.year,
-          seasonNumber: ep.season.seasonNumber,
-          episodeNumber: ep.episodeNumber,
-          episodeTitle: ep.title,
-          hasFile: ep.hasFile,
-        });
-      }
-    }
-
-    results.sort((a, b) => a.date.localeCompare(b.date));
-    return results;
+    return this.query.getCalendar(dto, accessibleLibraryIds, userId);
   }
 
-  // ---------------------------------------------------------------------------
-  // History
-  // ---------------------------------------------------------------------------
+  // -- Mutations --------------------------------------------------------------
 
-  async updateSeason(
+  update(id: number, dto: UpdateMediaDto) {
+    return this.mutation.update(id, dto);
+  }
+
+  updateLibrary(id: number, libraryId: number) {
+    return this.mutation.updateLibrary(id, libraryId);
+  }
+
+  updateProfiles(id: number, dto: UpdateMediaProfilesDto) {
+    return this.mutation.updateProfiles(id, dto);
+  }
+
+  bulkUpdate(dto: BulkUpdateMediaDto) {
+    return this.mutation.bulkUpdate(dto);
+  }
+
+  remove(id: number) {
+    return this.mutation.remove(id);
+  }
+
+  updateSeason(
     seasonId: number,
     patch: { monitored?: boolean; preferredProvider?: 'tmdb' | 'tvdb' | null },
-  ): Promise<Season> {
-    const season = await this.seasonRepo.findOne({
-      where: { id: seasonId },
-      relations: ['media'],
-    });
-    if (!season) throw new NotFoundException(`Season #${seasonId} not found`);
-    const wasMonitored = season.monitored;
-    if (patch.monitored !== undefined) season.monitored = patch.monitored;
-    if (patch.preferredProvider !== undefined)
-      season.preferredProvider = patch.preferredProvider;
-    const saved = await this.seasonRepo.save(season);
-    if (season.media) {
-      await this.requestLifecycle.onSeasonMonitorChange(
-        season.media,
-        saved.seasonNumber,
-        wasMonitored,
-        saved.monitored,
-      );
-    }
-    return saved;
+  ) {
+    return this.mutation.updateSeason(seasonId, patch);
   }
 
-  async updateEpisodeMonitored(
-    episodeId: number,
-    monitored: boolean,
-  ): Promise<Episode> {
-    const episode = await this.episodeRepo.findOne({
-      where: { id: episodeId },
-    });
-    if (!episode)
-      throw new NotFoundException(`Episode #${episodeId} not found`);
-    episode.monitored = monitored;
-    return this.episodeRepo.save(episode);
+  updateEpisodeMonitored(episodeId: number, monitored: boolean) {
+    return this.mutation.updateEpisodeMonitored(episodeId, monitored);
   }
 
-  async deleteMediaFile(
-    mediaId: number,
-    fileId: number,
-    deleteOnDisk: boolean,
-  ): Promise<void> {
-    const media = await this.mediaRepo.findOne({ where: { id: mediaId } });
-    if (!media) throw new NotFoundException(`Media #${mediaId} not found`);
-
-    const file = await this.mediaFileRepo.findOne({
-      where: { id: fileId, media: { id: mediaId } },
-    });
-    if (!file) throw new NotFoundException(`File #${fileId} not found`);
-
-    if (deleteOnDisk && media.path) {
-      const fs = await import('fs');
-      const path = await import('path');
-      const fullPath = path.join(media.path, file.relativePath);
-      try {
-        fs.unlinkSync(fullPath);
-        this.log.log(`Deleted file on disk: ${fullPath}`);
-      } catch (err: unknown) {
-        const code = (err as NodeJS.ErrnoException).code;
-        if (code !== 'ENOENT') throw err;
-        this.log.warn(`File not found on disk (already deleted?): ${fullPath}`);
-      }
-    }
-
-    const episodeId = file.episodeId;
-    await this.mediaFileRepo.remove(file);
-    if (episodeId != null) {
-      const remaining = await this.mediaFileRepo.count({
-        where: { episode: { id: episodeId } },
-      });
-      if (remaining === 0) {
-        await this.episodeRepo.update(episodeId, { hasFile: false });
-      }
-    }
-
-    void this.mediaServers.dispatch('file.deleted', {
-      title: media.title,
-      path: media.path,
-    });
+  deleteMediaFile(mediaId: number, fileId: number, deleteOnDisk: boolean) {
+    return this.mutation.deleteMediaFile(mediaId, fileId, deleteOnDisk);
   }
 
-  async refreshMetadata(id: number): Promise<Media> {
-    const media = await this.mediaRepo.findOne({ where: { id } });
-    if (!media) throw new NotFoundException(`Media #${id} not found`);
+  // -- Metadata ---------------------------------------------------------------
 
-    const { provider, externalId } = await this.resolveProviderForMedia(media);
-
-    this.log.log(
-      `refreshMetadata: "${media.title}" using provider=${provider.name} externalId=${externalId}`,
-    );
-
-    if (media.type === MediaType.MOVIE) {
-      const details = await provider.getMovieDetails(externalId);
-      // Cross-ref to fill missing IDs
-      if (!media.tvdbId && details.tvdbId)
-        await this.mediaRepo.update(media.id, { tvdbId: details.tvdbId });
-      if (!media.tmdbId && details.tmdbId)
-        await this.mediaRepo.update(media.id, { tmdbId: details.tmdbId });
-      await this.mediaRepo.update(media.id, {
-        ...this.buildMediaFieldsFromTmdb(details, MediaType.MOVIE),
-      });
-      await this.downloadMediaImages(media.id, details);
-      await this.persistMediaMetadata(media, details);
-    } else {
-      const details = await provider.getTvShowDetails(externalId);
-      if (!media.tvdbId && details.tvdbId)
-        await this.mediaRepo.update(media.id, { tvdbId: details.tvdbId });
-      if (!media.tmdbId && details.tmdbId)
-        await this.mediaRepo.update(media.id, { tmdbId: details.tmdbId });
-      await this.mediaRepo.update(media.id, {
-        ...this.buildMediaFieldsFromTmdb(details, MediaType.SERIES),
-      });
-      await this.downloadMediaImages(media.id, details);
-      await this.persistMediaMetadata(media, details);
-      await this.refreshSeriesEpisodes(media, { provider, externalId });
-    }
-
-    await this.updateSearchVector(media.id);
-
-    await this.mediaRepo.update(media.id, {
-      metadataRefreshedAt: new Date(),
-    });
-
-    return this.findOne(media.id);
+  refreshMetadata(id: number) {
+    return this.metadata.refreshMetadata(id);
   }
 
-  async refreshEpisodeMetadata(
-    mediaId: number,
-    episodeId: number,
-  ): Promise<Media> {
-    const episode = await this.episodeRepo.findOne({
-      where: { id: episodeId },
-      relations: ['season'],
-    });
-    if (!episode)
-      throw new NotFoundException(`Episode #${episodeId} not found`);
-
-    const media = await this.mediaRepo.findOne({ where: { id: mediaId } });
-    if (!media) throw new NotFoundException(`Media #${mediaId} not found`);
-
-    const { provider, externalId } = await this.resolveProviderForMedia(media, {
-      season: episode.season,
-    });
-
-    const seasonData = await this.fetchSingleSeason(
-      provider,
-      externalId,
-      episode.season.seasonNumber,
-    );
-    const tmdbEp = seasonData?.episodes.find(
-      (e) => e.episodeNumber === episode.episodeNumber,
-    );
-
-    if (tmdbEp) {
-      const updates: Partial<Episode> = {};
-      if (tmdbEp.title && tmdbEp.title !== episode.title)
-        updates.title = tmdbEp.title;
-      if (tmdbEp.overview && tmdbEp.overview !== episode.overview)
-        updates.overview = tmdbEp.overview;
-      if (tmdbEp.airDate && tmdbEp.airDate !== episode.airDate)
-        updates.airDate = tmdbEp.airDate;
-      if (tmdbEp.runtime != null && tmdbEp.runtime !== episode.runtime)
-        updates.runtime = tmdbEp.runtime;
-      if (Object.keys(updates).length > 0) {
-        await this.episodeRepo.update(episode.id, updates);
-      }
-      if (tmdbEp.stillUrl) {
-        await this.downloadEpisodeStill(episode.id, tmdbEp.stillUrl);
-      }
-    }
-
-    return this.findOne(mediaId);
+  async refreshEpisodeMetadata(mediaId: number, episodeId: number) {
+    await this.metadata.refreshEpisodeMetadata(mediaId, episodeId);
+    return this.query.findOne(mediaId);
   }
 
-  private async refreshSeriesEpisodes(
+  // -- Rescan + analyse + file enrichment -------------------------------------
+
+  rescanFiles(mediaId: number, options?: { skipWarmup?: boolean }) {
+    return this.rescan.rescanFiles(mediaId, options);
+  }
+
+  analyzeMedia(mediaId: number, opts: AnalyzeMediaDto) {
+    return this.rescan.analyzeMedia(mediaId, opts);
+  }
+
+  enrichMediaFileFromDisk(mediaFileId: number) {
+    return this.rescan.enrichMediaFileFromDisk(mediaFileId);
+  }
+
+  finalizeImportedFile(
+    file: import('./entities/media-file.entity').MediaFile,
+    absPath: string,
     media: Media,
-    preResolved?: { provider: IMetadataProvider; externalId: string },
-  ): Promise<void> {
-    // 1. Media-level provider enumerates all seasons and provides defaults.
-    //    `refreshMetadata` has already resolved the provider — accept it
-    //    so we don't re-emit the `resolveProvider:` log and skip the DB
-    //    lookup. Cron / non-refreshMetadata callers omit `preResolved`
-    //    and we resolve on their behalf.
-    const mediaResolve =
-      preResolved ?? (await this.resolveProviderForMedia(media));
-    const mediaSeasons = await mediaResolve.provider.getTvShowSeasons(
-      mediaResolve.externalId,
-    );
-    const dbSeasons = await this.seasonRepo.find({
-      where: { media: { id: media.id } },
-      relations: ['episodes'],
-    });
-    const dbSeasonMap = new Map(dbSeasons.map((s) => [s.seasonNumber, s]));
-
-    // 2. Upsert seasons + apply media-level episode data, skipping seasons
-    //    whose override points elsewhere (handled in the second pass).
-    for (const sd of mediaSeasons) {
-      let dbSeason = dbSeasonMap.get(sd.seasonNumber);
-      if (!dbSeason) {
-        dbSeason = await this.seasonRepo.save(
-          this.seasonRepo.create({
-            media,
-            seasonNumber: sd.seasonNumber,
-            monitored: true,
-          }),
-        );
-        dbSeason.episodes = [];
-        dbSeasonMap.set(sd.seasonNumber, dbSeason);
-      }
-      if (
-        dbSeason.preferredProvider &&
-        dbSeason.preferredProvider !== mediaResolve.provider.name
-      ) {
-        continue;
-      }
-      await this.applySeasonDetails(dbSeason, sd);
-    }
-
-    // 3. Second pass: re-fetch overridden seasons from their own provider.
-    //    Cached by (providerName, externalId) so N overrides to the same
-    //    non-TMDB provider only trigger one getTvShowSeasons call.
-    const overridden = dbSeasons.filter(
-      (s) =>
-        s.preferredProvider &&
-        s.preferredProvider !== mediaResolve.provider.name,
-    );
-    if (!overridden.length) return;
-    const cache = new Map<string, SeasonDetails[]>();
-    for (const dbSeason of overridden) {
-      let overrideResolve: {
-        provider: IMetadataProvider;
-        externalId: string;
-      };
-      try {
-        overrideResolve = await this.resolveProviderForMedia(media, {
-          season: dbSeason,
-        });
-      } catch (err) {
-        this.log.warn(
-          `refreshSeriesEpisodes: season S${dbSeason.seasonNumber} override ` +
-            `"${dbSeason.preferredProvider}" failed — ${err instanceof Error ? err.message : 'unknown error'}. ` +
-            `Falling back to media-level data.`,
-        );
-        const fallback = mediaSeasons.find(
-          (s) => s.seasonNumber === dbSeason.seasonNumber,
-        );
-        if (fallback) await this.applySeasonDetails(dbSeason, fallback);
-        continue;
-      }
-      const seasonData = await this.fetchSingleSeason(
-        overrideResolve.provider,
-        overrideResolve.externalId,
-        dbSeason.seasonNumber,
-        cache,
-      );
-      if (!seasonData) {
-        this.log.warn(
-          `refreshSeriesEpisodes: season S${dbSeason.seasonNumber} override ` +
-            `"${dbSeason.preferredProvider}" returned no matching season ` +
-            `(numbering mismatch?). Falling back to media-level data.`,
-        );
-        const fallback = mediaSeasons.find(
-          (s) => s.seasonNumber === dbSeason.seasonNumber,
-        );
-        if (fallback) await this.applySeasonDetails(dbSeason, fallback);
-        continue;
-      }
-      await this.applySeasonDetails(dbSeason, seasonData);
-    }
+  ) {
+    return this.rescan.finalizeImportedFile(file, absPath, media);
   }
 
-  /** Upsert episodes of one DB season from provider season details. */
-  private async applySeasonDetails(
-    dbSeason: Season,
-    sd: SeasonDetails,
-  ): Promise<void> {
-    const dbEpMap = new Map(
-      (dbSeason.episodes ?? []).map((e) => [e.episodeNumber, e]),
-    );
-
-    // 1. Partition into INSERTs and UPDATEs in a single pass.
-    type UpdateJob = {
-      id: number;
-      updates: Partial<Episode>;
-      stillUrl?: string | null;
-    };
-    type InsertEntry = (typeof sd.episodes)[number];
-    const updateJobs: UpdateJob[] = [];
-    const inserts: InsertEntry[] = [];
-    for (const ep of sd.episodes) {
-      const existing = dbEpMap.get(ep.episodeNumber);
-      if (existing) {
-        const updates: Partial<Episode> = {};
-        if (ep.title && ep.title !== existing.title) updates.title = ep.title;
-        if (ep.overview && ep.overview !== existing.overview)
-          updates.overview = ep.overview;
-        if (ep.airDate && ep.airDate !== existing.airDate)
-          updates.airDate = ep.airDate;
-        if (ep.runtime != null && ep.runtime !== existing.runtime)
-          updates.runtime = ep.runtime;
-        if (ep.stillUrl || Object.keys(updates).length > 0) {
-          updateJobs.push({ id: existing.id, updates, stillUrl: ep.stillUrl });
-        }
-      } else {
-        inserts.push(ep);
-      }
-    }
-
-    // 2. Batch INSERT new episodes in one save() call.
-    const insertedRows = inserts.length
-      ? await this.episodeRepo.save(
-          inserts.map((ep) =>
-            this.episodeRepo.create({
-              season: dbSeason,
-              episodeNumber: ep.episodeNumber,
-              title: ep.title || undefined,
-              overview: ep.overview || undefined,
-              airDate: ep.airDate || undefined,
-              runtime: ep.runtime ?? undefined,
-              monitored: true,
-            }),
-          ),
-        )
-      : [];
-
-    // 3. Parallelize UPDATEs + still downloads (concurrency 8). On a long
-    //    season (~20 ep × 2 image variants) this collapses ~40 sequential
-    //    HTTPs into ~5 batches.
-    type Job = {
-      id: number;
-      updates?: Partial<Episode>;
-      stillUrl?: string | null;
-    };
-    const jobs: Job[] = [
-      ...updateJobs,
-      ...insertedRows.map((row, i) => ({
-        id: row.id,
-        stillUrl: inserts[i].stillUrl,
-      })),
-    ];
-    await mapWithConcurrency(jobs, 8, async ({ id, updates, stillUrl }) => {
-      if (updates && Object.keys(updates).length > 0) {
-        await this.episodeRepo.update(id, updates);
-      }
-      if (stillUrl) {
-        await this.downloadEpisodeStill(id, stillUrl);
-      }
-    });
-
-    if (sd.posterUrl) {
-      await this.downloadSeasonPoster(dbSeason.id, sd.posterUrl);
-    }
-  }
+  // -- Request-driven lookups -------------------------------------------------
 
   /**
-   * Fetch one season from the given provider. Uses TMDB's fast single-season
-   * endpoint when possible; otherwise falls back to a full `getTvShowSeasons`
-   * fetch (memoized via `cache` so sibling overrides don't re-fetch it).
+   * Load a media row with the relations the lifecycle needs to compute
+   * coverage (`files`, `seasons`, `seasons.episodes`). Centralised here
+   * to keep `mediaRepo.findOne` calls in one place.
    */
-  private async fetchSingleSeason(
-    provider: IMetadataProvider,
-    externalId: string,
-    seasonNumber: number,
-    cache?: Map<string, SeasonDetails[]>,
-  ): Promise<SeasonDetails | undefined> {
-    if (provider.name === 'tmdb') {
-      return this.tmdb.getTvSeason(externalId, seasonNumber);
-    }
-    const key = `${provider.name}:${externalId}`;
-    let all = cache?.get(key);
-    if (!all) {
-      all = await provider.getTvShowSeasons(externalId);
-      cache?.set(key, all);
-    }
-    return all.find((s) => s.seasonNumber === seasonNumber);
-  }
-
-  private applyFilters(
-    qb: SelectQueryBuilder<Media>,
-    query: SearchMediaDto,
-  ): void {
-    if (query.libraryId) {
-      qb.andWhere('media.libraryId = :libraryId', {
-        libraryId: query.libraryId,
-      });
-    }
-    if (query.type) {
-      qb.andWhere('media.type = :type', { type: query.type });
-    }
-    if (query.status) {
-      qb.andWhere('media.status = :status', { status: query.status });
-    }
-    if (query.monitored !== undefined) {
-      qb.andWhere('media.monitored = :monitored', {
-        monitored: query.monitored,
-      });
-    }
-    if (query.year) {
-      qb.andWhere('media.year = :year', { year: query.year });
-    }
-    if (query.genre) {
-      qb.andWhere('media.genres @> :genre', {
-        genre: JSON.stringify([query.genre]),
-      });
-    }
-    if (query.collectionId) {
-      qb.andWhere('media.tmdbCollectionId = :collectionId', {
-        collectionId: query.collectionId,
-      });
-    }
-    if (query.qualityProfileId) {
-      qb.andWhere('media.qualityProfileId = :qpId', {
-        qpId: query.qualityProfileId,
-      });
-    }
-    if (query.languageProfileId) {
-      qb.andWhere('media.languageProfileId = :lpId', {
-        lpId: query.languageProfileId,
-      });
-    }
-    if (query.missing === true) {
-      qb.andWhere('files.id IS NULL');
-    } else if (query.missing === false) {
-      qb.andWhere('files.id IS NOT NULL');
-    }
-    if (query.letter) {
-      const letter = query.letter.toUpperCase();
-      if (letter === '#') {
-        qb.andWhere(`media.title !~ '^[A-Za-z]'`);
-      } else if (/^[A-Z]$/.test(letter)) {
-        qb.andWhere(`UPPER(LEFT(media.title, 1)) = :letter`, { letter });
-      }
-    }
-  }
-
-  private applyFullTextSearch(
-    qb: SelectQueryBuilder<Media>,
-    searchTerm: string,
-  ): void {
-    qb.addSelect(
-      `ts_rank(media."searchVector", plainto_tsquery('french', :q))`,
-      'rank',
-    );
-    qb.andWhere(
-      `(
-        media."searchVector" @@ plainto_tsquery('french', :q)
-        OR media.title ILIKE :like
-        OR media."originalTitle" ILIKE :like
-        OR similarity(media.title, :q) > 0.8
-      )`,
-      { q: searchTerm, like: `%${searchTerm}%` },
-    );
-    qb.orderBy('rank', 'DESC');
-  }
-
-  private async updateSearchVector(mediaId: number): Promise<void> {
-    await this.dataSource.query(
-      `UPDATE media SET "searchVector" =
-        setweight(to_tsvector('french', COALESCE(title, '')), 'A') ||
-        setweight(to_tsvector('french', COALESCE("originalTitle", '')), 'B') ||
-        setweight(to_tsvector('french', COALESCE(overview, '')), 'C')
-      WHERE id = $1`,
-      [mediaId],
-    );
-  }
-
-  private mapTmdbStatusToMediaStatus(
-    type: MediaType,
-    status: string,
-  ): MediaStatus {
-    const s = (status || '').toLowerCase();
-    if (type === MediaType.MOVIE) {
-      const m: Record<string, MediaStatus> = {
-        released: MediaStatus.RELEASED,
-        rumored: MediaStatus.TBA,
-        rumor: MediaStatus.TBA,
-        planned: MediaStatus.ANNOUNCED,
-        'in production': MediaStatus.ANNOUNCED,
-        'post production': MediaStatus.ANNOUNCED,
-        canceled: MediaStatus.ENDED,
-        cancelled: MediaStatus.ENDED,
-      };
-      return m[s] ?? MediaStatus.TBA;
-    }
-    const m: Record<string, MediaStatus> = {
-      continuing: MediaStatus.CONTINUING,
-      ended: MediaStatus.ENDED,
-      announced: MediaStatus.ANNOUNCED,
-      tba: MediaStatus.TBA,
-      unknown: MediaStatus.TBA,
-    };
-    return m[s] ?? MediaStatus.TBA;
-  }
-
-  /**
-   * Resolve which metadata provider + external ID to use for a media.
-   *
-   * Priority: season override → media override → library preference → fallback.
-   *
-   * Explicit overrides (season, media) are strict: if the preferred provider
-   * is configured but its external ID can't be resolved (missing cross-ref),
-   * this throws instead of silently falling back — so a user's "use TVDB for
-   * this" choice never degrades to TMDB without them knowing. The library-
-   * level preference stays soft (logs a warning, falls back) since it's a
-   * hint across many medias rather than a per-item contract.
-   */
-  private async resolveProviderForMedia(
-    media: Media,
-    opts?: { season?: Season },
-  ): Promise<{ provider: IMetadataProvider; externalId: string }> {
-    const label = `"${media.title}" (#${media.id})`;
-
-    // 1. Season override (strict)
-    const seasonPref = opts?.season?.preferredProvider;
-    if (seasonPref) {
-      const resolved = await this.tryPreferred(media, seasonPref, 'season');
-      if (resolved) return resolved;
-      throw new BadRequestException(
-        `Season override "${seasonPref}" cannot be resolved for ${label}: ` +
-          `no matching external ID available (check API key + cross-reference).`,
-      );
-    }
-
-    // 2. Media override (strict)
-    if (media.preferredProvider) {
-      const resolved = await this.tryPreferred(
-        media,
-        media.preferredProvider,
-        'media',
-      );
-      if (resolved) return resolved;
-      throw new BadRequestException(
-        `Media override "${media.preferredProvider}" cannot be resolved for ` +
-          `${label}: no matching external ID available.`,
-      );
-    }
-
-    // 3. Library preference (soft)
-    if (media.libraryId) {
-      const lib = await this.libraryRepo.findOne({
-        where: { id: media.libraryId },
-      });
-      const pref = lib?.preferredProvider;
-      if (pref) {
-        const resolved = await this.tryPreferred(media, pref, 'library');
-        if (resolved) return resolved;
-        this.log.warn(
-          `resolveProvider: ${label} — library preferred ${pref} but no matching ID found, falling back`,
-        );
-      }
-    }
-
-    // 4. Fallback: use whichever ID + provider is available
-    if (media.tmdbId && this.providerRegistry.isAvailable('tmdb')) {
-      this.log.log(
-        `resolveProvider: ${label} — fallback to tmdb (tmdbId=${media.tmdbId})`,
-      );
-      return { provider: this.tmdb, externalId: String(media.tmdbId) };
-    }
-    if (media.tvdbId && this.providerRegistry.isAvailable('tvdb')) {
-      this.log.log(
-        `resolveProvider: ${label} — fallback to tvdb (tvdbId=${media.tvdbId})`,
-      );
-      return {
-        provider: this.providerRegistry.get('tvdb')!,
-        externalId: String(media.tvdbId),
-      };
-    }
-    if (media.tmdbId) {
-      this.log.log(
-        `resolveProvider: ${label} — fallback to tmdb (tmdbId=${media.tmdbId}, unchecked availability)`,
-      );
-      return { provider: this.tmdb, externalId: String(media.tmdbId) };
-    }
-    throw new BadRequestException('No provider ID available for this media');
-  }
-
-  /**
-   * Try to resolve a preferred provider by name. Returns null if the provider
-   * is not configured (missing API key) or if cross-referencing to its ID
-   * fails. Caller decides whether to throw (strict) or fall through (soft).
-   */
-  private async tryPreferred(
-    media: Media,
-    pref: string,
-    origin: 'season' | 'media' | 'library',
-  ): Promise<{ provider: IMetadataProvider; externalId: string } | null> {
-    const label = `"${media.title}" (#${media.id})`;
-    this.log.log(`resolveProvider: ${label} — ${origin} prefers ${pref}`);
-    if (!this.providerRegistry.isAvailable(pref)) {
-      this.log.warn(
-        `resolveProvider: ${label} — ${origin} preferred ${pref} but not available (no API key?)`,
-      );
-      return null;
-    }
-    const provider = this.providerRegistry.get(pref)!;
-    const externalId = await this.resolveExternalIdForProvider(
-      media,
-      provider,
-      pref,
-    );
-    if (!externalId) return null;
-    this.log.log(
-      `resolveProvider: ${label} — using ${pref} with id=${externalId} (${origin})`,
-    );
-    return { provider, externalId };
-  }
-
-  /**
-   * Find the external ID for a given provider on a media.
-   * If the media doesn't have the matching ID, cross-reference via other providers.
-   */
-  private async resolveExternalIdForProvider(
-    media: Media,
-    provider: IMetadataProvider,
-    providerName: string,
-  ): Promise<string | null> {
-    const label = `"${media.title}" (#${media.id})`;
-
-    // Direct match
-    if (providerName === 'tmdb' && media.tmdbId) return String(media.tmdbId);
-    if (providerName === 'tvdb' && media.tvdbId) return String(media.tvdbId);
-
-    this.log.log(
-      `crossRef: ${label} — need ${providerName} ID, attempting cross-reference`,
-    );
-
-    // Cross-reference: need to find the missing ID
-    if (providerName === 'tvdb' && !media.tvdbId) {
-      if (media.imdbId && provider.findByExternalId) {
-        this.log.log(
-          `crossRef: ${label} — trying TVDB lookup via imdbId=${media.imdbId}`,
-        );
-        const cross = await provider.findByExternalId('imdb', media.imdbId);
-        if (cross) {
-          this.log.log(
-            `crossRef: ${label} — found tvdbId=${cross.id} via IMDB`,
-          );
-          await this.mediaRepo.update(media.id, {
-            tvdbId: parseInt(cross.id, 10),
-          });
-          return cross.id;
-        }
-      }
-      if (media.tmdbId) {
-        this.log.log(
-          `crossRef: ${label} — trying TMDB external_ids for tvdbId (tmdbId=${media.tmdbId})`,
-        );
-        const details =
-          media.type === MediaType.MOVIE
-            ? await this.tmdb.getMovieDetails(String(media.tmdbId))
-            : await this.tmdb.getTvShowDetails(String(media.tmdbId));
-        if (details.tvdbId) {
-          this.log.log(
-            `crossRef: ${label} — found tvdbId=${details.tvdbId} via TMDB`,
-          );
-          await this.mediaRepo.update(media.id, { tvdbId: details.tvdbId });
-          return String(details.tvdbId);
-        }
-      }
-    }
-
-    if (providerName === 'tmdb' && !media.tmdbId) {
-      if (media.imdbId && this.tmdb.findByExternalId) {
-        this.log.log(
-          `crossRef: ${label} — trying TMDB find via imdbId=${media.imdbId}`,
-        );
-        const cross = await this.tmdb.findByExternalId('imdb', media.imdbId);
-        if (cross) {
-          this.log.log(
-            `crossRef: ${label} — found tmdbId=${cross.id} via IMDB`,
-          );
-          await this.mediaRepo.update(media.id, {
-            tmdbId: parseInt(cross.id, 10),
-          });
-          return cross.id;
-        }
-      }
-      if (media.tvdbId && this.tmdb.findByExternalId) {
-        this.log.log(
-          `crossRef: ${label} — trying TMDB find via tvdbId=${media.tvdbId}`,
-        );
-        const cross = await this.tmdb.findByExternalId(
-          'tvdb',
-          String(media.tvdbId),
-        );
-        if (cross) {
-          this.log.log(
-            `crossRef: ${label} — found tmdbId=${cross.id} via TVDB`,
-          );
-          await this.mediaRepo.update(media.id, {
-            tmdbId: parseInt(cross.id, 10),
-          });
-          return cross.id;
-        }
-      }
-    }
-
-    this.log.warn(
-      `crossRef: ${label} — cross-reference failed for ${providerName}`,
-    );
-    return null;
-  }
-
-  private buildMediaFieldsFromTmdb(
-    details: MetadataDetails,
-    type: MediaType,
-  ): Partial<Media> {
-    const year =
-      details.year != null && Number.isFinite(details.year)
-        ? details.year
-        : undefined;
-    return {
-      title: details.title,
-      originalTitle: details.originalTitle ?? details.title,
-      alternativeTitles: details.alternativeTitles ?? [],
-      year,
-      type,
-      tmdbId: details.tmdbId || undefined,
-      tvdbId: details.tvdbId ?? undefined,
-      imdbId: details.imdbId ?? undefined,
-      overview: details.overview ?? undefined,
-      status: this.mapTmdbStatusToMediaStatus(type, details.status),
-      posterUrl: details.posterUrl ?? undefined,
-      fanartUrl: details.fanartUrl ?? undefined,
-      rating: details.rating ?? undefined,
-      genres: details.genres?.length ? details.genres : [],
-      runtime: details.runtime ?? undefined,
-      releaseDate: details.releaseDate
-        ? details.releaseDate.slice(0, 10)
-        : undefined,
-      inCinemas: details.inCinemas ? details.inCinemas.slice(0, 10) : undefined,
-      digitalRelease: details.digitalRelease
-        ? details.digitalRelease.slice(0, 10)
-        : undefined,
-      physicalRelease: details.physicalRelease
-        ? details.physicalRelease.slice(0, 10)
-        : undefined,
-      tmdbCollectionId: details.tmdbCollectionId ?? null,
-      tmdbCollectionName: details.tmdbCollectionName ?? null,
-    };
-  }
-
-  /**
-   * Download poster + fanart from TMDB and update the media row with local paths.
-   */
-  private async downloadMediaImages(
-    mediaId: number,
-    details: MetadataDetails,
-  ): Promise<void> {
-    const updates: Partial<Media> = {};
-
-    if (details.posterUrl) {
-      const local = await this.imageService.downloadAndStore(
-        details.posterUrl,
-        'media',
-        mediaId,
-        'poster',
-      );
-      if (local) updates.posterUrl = local;
-    }
-    if (details.fanartUrl) {
-      const local = await this.imageService.downloadAndStore(
-        details.fanartUrl,
-        'media',
-        mediaId,
-        'fanart',
-      );
-      if (local) updates.fanartUrl = local;
-    }
-
-    // Extra fanarts (variants fanart-1..N). Downloaded in parallel
-    // since each entry is an independent CDN GET. Slots are
-    // 1-indexed so the on-disk filename mirrors the variant string
-    // a frontend caller passes back (`/api/images/media/X/fanart-1`).
-    if (details.additionalFanartUrls?.length) {
-      const downloaded = await Promise.all(
-        details.additionalFanartUrls.map((url: string, i: number) =>
-          this.imageService.downloadAndStore(
-            url,
-            'media',
-            mediaId,
-            `fanart-${i + 1}`,
-          ),
-        ),
-      );
-      updates.additionalFanartUrls = downloaded.filter(
-        (p): p is string => !!p,
-      );
-    } else {
-      // Provider returned no extras → clear stale entries from a
-      // previous refresh that did. Keeps the column consistent
-      // with provider reality.
-      updates.additionalFanartUrls = [];
-    }
-
-    if (Object.keys(updates).length > 0) {
-      await this.mediaRepo.update(mediaId, updates);
-    }
-  }
-
-  /**
-   * Download a person avatar from TMDB and update the person row.
-   */
-  private async downloadPersonAvatar(
-    personId: number,
-    avatarUrl: string,
-  ): Promise<string | null> {
-    const local = await this.imageService.downloadAndStore(
-      avatarUrl,
-      'person',
-      personId,
-    );
-    if (local) {
-      await this.personRepo.update(personId, { avatarUrl: local });
-    }
-    return local;
-  }
-
-  /**
-   * Download an episode still from TMDB and update the episode row.
-   */
-  private async downloadEpisodeStill(
-    episodeId: number,
-    stillUrl: string,
-  ): Promise<void> {
-    const local = await this.imageService.downloadAndStore(
-      stillUrl,
-      'episode',
-      episodeId,
-    );
-    if (local) {
-      await this.episodeRepo.update(episodeId, { stillUrl: local });
-    }
-  }
-
-  /**
-   * Download a season poster from TMDB / TVDB and persist the local
-   * API path on the Season row. Best-effort: a failed download leaves
-   * `posterUrl` null on the DB row and the UI falls back to the
-   * series poster.
-   */
-  private async downloadSeasonPoster(
-    seasonId: number,
-    posterUrl: string,
-  ): Promise<void> {
-    const local = await this.imageService.downloadAndStore(
-      posterUrl,
-      'season',
-      seasonId,
-    );
-    if (local) {
-      await this.seasonRepo.update(seasonId, { posterUrl: local });
-    }
-  }
-
-  private async persistMediaMetadata(
-    media: Media,
-    details: MetadataDetails,
-  ): Promise<void> {
-    // Upsert MediaMetadata
-    const existing = await this.metadataRepo.findOne({
-      where: { media: { id: media.id } },
-    });
-    const metaFields = {
-      budget: details.budget ?? undefined,
-      revenue: details.revenue ?? undefined,
-      tagline: details.tagline ?? undefined,
-      popularity: details.popularity ?? undefined,
-      voteCount: details.voteCount ?? undefined,
-      originalLanguage: details.originalLanguage ?? undefined,
-      productionCountries: details.productionCountries,
-      productionCompanies: details.productionCompanies,
-      videos: details.videos,
-      keywords: details.keywords,
-    };
-    if (existing) {
-      await this.metadataRepo.update(existing.id, metaFields);
-    } else {
-      await this.metadataRepo.save(
-        this.metadataRepo.create({ media, ...metaFields } as MediaMetadata),
-      );
-    }
-
-    // Upsert Persons + replace cast/crew
-    const personMap = new Map<number, Person>();
-    const allExternalIds = [
-      ...details.cast.map((c) => c.externalId),
-      ...details.crew.map((c) => c.externalId),
-    ];
-    const uniqueIds = [...new Set(allExternalIds)];
-
-    if (uniqueIds.length > 0) {
-      const existingPersons = await this.personRepo
-        .createQueryBuilder('p')
-        .where('p.tmdbId IN (:...ids)', { ids: uniqueIds })
-        .getMany();
-      for (const p of existingPersons) personMap.set(p.tmdbId, p);
-
-      const allCredits = [...details.cast, ...details.crew];
-
-      // Batch INSERT new persons in a single save() call instead of one
-      // round-trip per missing person.
-      const missingIds = uniqueIds.filter((id) => !personMap.has(id));
-      const newRows = missingIds
-        .map((id) => allCredits.find((c) => c.externalId === id))
-        .filter((c): c is (typeof allCredits)[number] => !!c)
-        .map((c) =>
-          this.personRepo.create({ tmdbId: c.externalId, name: c.name }),
-        );
-      const inserted = newRows.length
-        ? await this.personRepo.save(newRows)
-        : [];
-      for (const p of inserted) personMap.set(p.tmdbId, p);
-
-      // Parallelize avatar downloads + name updates across new AND existing
-      // persons (concurrency 8). Each job: download avatar (best-effort),
-      // then UPDATE the row with whatever fields actually changed. This is
-      // the dominant cost on series with large casts.
-      type PersonJob = {
-        id: number;
-        updates: Partial<Person>;
-        avatarUrl: string | null;
-      };
-      const jobs: PersonJob[] = [];
-      for (const p of inserted) {
-        const credit = allCredits.find((c) => c.externalId === p.tmdbId);
-        if (credit?.avatarUrl) {
-          jobs.push({ id: p.id, updates: {}, avatarUrl: credit.avatarUrl });
-        }
-      }
-      for (const p of existingPersons) {
-        const credit = allCredits.find((c) => c.externalId === p.tmdbId);
-        if (!credit) continue;
-        const updates: Partial<Person> = {};
-        if (credit.name !== p.name) updates.name = credit.name;
-        if (credit.avatarUrl || Object.keys(updates).length > 0) {
-          jobs.push({ id: p.id, updates, avatarUrl: credit.avatarUrl ?? null });
-        }
-      }
-      await mapWithConcurrency(jobs, 8, async ({ id, updates, avatarUrl }) => {
-        const final = { ...updates };
-        if (avatarUrl) {
-          const local = await this.imageService.downloadAndStore(
-            avatarUrl,
-            'person',
-            id,
-          );
-          if (local) final.avatarUrl = local;
-        }
-        if (Object.keys(final).length > 0) {
-          await this.personRepo.update(id, final);
-        }
-      });
-    }
-
-    // Replace cast + crew atomically — without a transaction, a refresh that
-    // crashes between DELETE and INSERT would leave the media with an empty
-    // cast/crew section in the UI until the next successful refresh.
-    await this.dataSource.transaction(async (manager) => {
-      const castMgr = manager.getRepository(MediaCast);
-      await castMgr.delete({ media: { id: media.id } });
-      if (details.cast.length > 0) {
-        await castMgr.insert(
-          details.cast.map((c) => ({
-            media: { id: media.id },
-            person: { id: personMap.get(c.externalId)?.id },
-            // TMDB returns null character for some cast entries (archival
-            // footage, uncredited roles). Columns are NOT NULL — coerce.
-            character: c.character ?? '',
-            order: c.order ?? 0,
-          })),
-        );
-      }
-      const crewMgr = manager.getRepository(MediaCrew);
-      await crewMgr.delete({ media: { id: media.id } });
-      if (details.crew.length > 0) {
-        await crewMgr.insert(
-          details.crew.map((c) => ({
-            media: { id: media.id },
-            person: { id: personMap.get(c.externalId)?.id },
-            job: c.job ?? '',
-            department: c.department ?? '',
-          })),
-        );
-      }
-    });
-
-    // Update search vectors + departments for persons
-    if (uniqueIds.length > 0) {
-      const personIds = [...personMap.values()].map((p) => p.id);
-      await this.dataSource.query(
-        `UPDATE persons SET "searchVector" = to_tsvector('simple', name) WHERE id = ANY($1)`,
-        [personIds],
-      );
-
-      // Compute departments from current details and merge with existing
-      const deptMap = new Map<number, Set<string>>();
-      for (const c of details.cast) {
-        if (!deptMap.has(c.externalId)) deptMap.set(c.externalId, new Set());
-        deptMap.get(c.externalId)!.add('Acting');
-      }
-      for (const c of details.crew) {
-        if (!deptMap.has(c.externalId)) deptMap.set(c.externalId, new Set());
-        deptMap.get(c.externalId)!.add(c.department);
-      }
-      for (const [tmdbId, person] of personMap) {
-        const newDepts = deptMap.get(tmdbId);
-        if (!newDepts) continue;
-        const merged = new Set(person.departments ?? []);
-        for (const d of newDepts) merged.add(d);
-        const sorted = [...merged].sort();
-        await this.personRepo.update(person.id, { departments: sorted });
-      }
-    }
-
-    // Refresh stale person details (biography, birthday, etc.)
-    const refreshThreshold = Date.now() - 7 * 24 * 60 * 60 * 1000;
-    for (const person of personMap.values()) {
-      const needsRefresh =
-        !person.metadataRefreshedAt ||
-        person.metadataRefreshedAt.getTime() < refreshThreshold;
-      if (!needsRefresh) continue;
-      try {
-        const pd = await this.tmdb.getPersonDetails(String(person.tmdbId));
-        let localAvatar: string | undefined;
-        if (pd.avatarUrl) {
-          const dl = await this.downloadPersonAvatar(person.id, pd.avatarUrl);
-          if (dl) localAvatar = dl;
-        }
-        await this.personRepo.update(person.id, {
-          name: pd.name,
-          ...(localAvatar ? { avatarUrl: localAvatar } : {}),
-          biography: pd.biography,
-          birthday: pd.birthday ?? undefined,
-          deathday: pd.deathday ?? undefined,
-          placeOfBirth: pd.placeOfBirth ?? undefined,
-          knownForDepartment: pd.knownForDepartment,
-          metadataRefreshedAt: new Date(),
-        });
-      } catch {
-        // Skip failed person detail fetches
-      }
-    }
-  }
-
-  private async persistImportedMovie(
-    details: MetadataDetails,
-    qualityProfileId: number | null,
-    languageProfileId: number | null,
-    folderName?: string,
-    libraryId?: number,
-    addedByUserId?: number | null,
-  ): Promise<Media> {
-    const row = this.mediaRepo.create({
-      ...this.buildMediaFieldsFromTmdb(details, MediaType.MOVIE),
-      monitored: true,
-      metadataRefreshedAt: new Date(),
-      ...(qualityProfileId != null
-        ? { qualityProfile: { id: qualityProfileId } as QualityProfile }
-        : {}),
-      ...(languageProfileId != null
-        ? { languageProfile: { id: languageProfileId } as LanguageProfile }
-        : {}),
-      ...(libraryId ? { library: { id: libraryId } as Library } : {}),
-      ...(folderName ? { folderName } : {}),
-      ...(addedByUserId ? { addedBy: { id: addedByUserId } as User } : {}),
-    });
-    const saved = await this.mediaRepo.save(row);
-    this.logLibraryAdded('movie', saved);
-    await this.downloadMediaImages(saved.id, details);
-    await this.updateSearchVector(saved.id);
-    await this.persistMediaMetadata(saved, details);
-    await this.requestLifecycle.onMediaImported(saved, addedByUserId ?? null);
-    return this.findOne(saved.id);
-  }
-
-  private logLibraryAdded(
-    kind: 'movie' | 'series',
-    media: Media,
-    extra?: string,
-  ): void {
-    const year = media.year ? ` (${media.year})` : '';
-    const tail = extra ? `, ${extra}` : '';
-    this.log.log(
-      `Library: added ${kind} "${media.title}"${year} — id=${media.id}, tmdbId=${media.tmdbId ?? '?'}${tail}`,
-    );
-  }
-
-  private async persistImportedSeries(
-    details: MetadataDetails,
-    seasons: SeasonDetails[],
-    qualityProfileId: number | null,
-    languageProfileId: number | null,
-    folderName?: string,
-    libraryId?: number,
-    addedByUserId?: number | null,
-  ): Promise<Media> {
-    const row = this.mediaRepo.create({
-      ...this.buildMediaFieldsFromTmdb(details, MediaType.SERIES),
-      monitored: true,
-      metadataRefreshedAt: new Date(),
-      ...(qualityProfileId != null
-        ? { qualityProfile: { id: qualityProfileId } as QualityProfile }
-        : {}),
-      ...(languageProfileId != null
-        ? { languageProfile: { id: languageProfileId } as LanguageProfile }
-        : {}),
-      ...(libraryId ? { library: { id: libraryId } as Library } : {}),
-      ...(folderName ? { folderName } : {}),
-      ...(addedByUserId ? { addedBy: { id: addedByUserId } as User } : {}),
-    });
-    const saved = await this.mediaRepo.save(row);
-    this.logLibraryAdded('series', saved, `seasons=${seasons.length}`);
-    await this.downloadMediaImages(saved.id, details);
-
-    for (const sd of seasons) {
-      const season = this.seasonRepo.create({
-        media: saved,
-        seasonNumber: sd.seasonNumber,
-        monitored: true,
-      });
-      const sSaved = await this.seasonRepo.save(season);
-      if (sd.episodes.length > 0) {
-        await this.episodeRepo.insert(
-          sd.episodes.map((ep) => ({
-            season: sSaved,
-            episodeNumber: ep.episodeNumber,
-            title: ep.title || undefined,
-            overview: ep.overview || undefined,
-            airDate: ep.airDate || undefined,
-            runtime: ep.runtime ?? undefined,
-            monitored: true,
-          })),
-        );
-      }
-      if (sd.posterUrl) {
-        await this.downloadSeasonPoster(sSaved.id, sd.posterUrl);
-      }
-    }
-
-    await this.updateSearchVector(saved.id);
-    await this.persistMediaMetadata(saved, details);
-    await this.requestLifecycle.onMediaImported(saved, addedByUserId ?? null);
-    return this.findOne(saved.id);
-  }
-
-  // ---------------------------------------------------------------------------
-  // Request-driven media operations — narrow public surface that
-  // `RequestLifecycleService` calls when it needs to mutate the media
-  // layer. The orchestration itself lives in the lifecycle service.
-  // ---------------------------------------------------------------------------
-
-  /** Load a media row with the relations the lifecycle needs to compute
-   *  coverage (`files`, `seasons`, `seasons.episodes`). Centralised here
-   *  to keep `mediaRepo.findOne` calls in one place. */
   findOneWithSeasonsAndFiles(mediaId: number): Promise<Media | null> {
     return this.mediaRepo.findOne({
       where: { id: mediaId },
@@ -2302,7 +228,11 @@ export class MediaService {
       await this.mediaRepo.save(media);
     }
     if (media.type !== MediaType.SERIES || !seasons?.length) return;
-    const rows = await this.seasonRepo.find({
+    // Sub-services don't own the Season repository directly here; load via
+    // the shared connection. We keep this orchestration on the façade since
+    // it spans two repos and is only ever called by RequestLifecycleService.
+    const seasonRepo = this.mediaRepo.manager.getRepository(Season);
+    const rows = await seasonRepo.find({
       where: { media: { id: media.id } },
     });
     const targets = rows.filter(
@@ -2310,834 +240,7 @@ export class MediaService {
     );
     if (targets.length) {
       for (const s of targets) s.monitored = true;
-      await this.seasonRepo.save(targets);
+      await seasonRepo.save(targets);
     }
-  }
-
-  /**
-   * Runs ffprobe on a file already stored in DB (e.g. disk import): streamInfo,
-   * crop, resolution-based quality — same path as rescan / download import.
-   */
-  async enrichMediaFileFromDisk(mediaFileId: number): Promise<void> {
-    const dbFile = await this.mediaFileRepo.findOne({
-      where: { id: mediaFileId },
-      relations: ['media'],
-    });
-    if (!dbFile?.media?.path) {
-      this.log.warn(
-        `enrichMediaFileFromDisk: file #${mediaFileId} missing or media has no path`,
-      );
-      return;
-    }
-    const mediaDir = path.resolve(dbFile.media.path);
-    const normPath = dbFile.relativePath?.replace(/\\/g, '/');
-    if (!normPath) return;
-    const absPath = path.join(mediaDir, normPath);
-    if (!fs.existsSync(absPath)) {
-      this.log.warn(`enrichMediaFileFromDisk: file not on disk — "${absPath}"`);
-      return;
-    }
-
-    let diskSize: number;
-    try {
-      diskSize = fs.statSync(absPath).size;
-    } catch (err) {
-      this.log.warn(
-        `enrichMediaFileFromDisk: cannot stat "${absPath}"`,
-        err instanceof Error ? err.stack : err,
-      );
-      return;
-    }
-
-    const filename = path.basename(absPath);
-
-    let streamInfo: Awaited<ReturnType<FfprobeService['detectMediaFileInfo']>>;
-    try {
-      streamInfo = await this.ffprobe.detectMediaFileInfo(absPath);
-    } catch (err) {
-      this.log.error(
-        `enrichMediaFileFromDisk: ffprobe failed for "${absPath}"`,
-        err instanceof Error ? err.stack : err,
-      );
-      return;
-    }
-
-    const qualityName = this.resolveQuality(
-      filename,
-      streamInfo.video?.[0]?.height,
-      streamInfo.video?.[0]?.width,
-    );
-    dbFile.size = diskSize;
-    dbFile.streamInfo = streamInfo;
-    dbFile.quality = qualityName;
-    const saved = await this.mediaFileRepo.save(dbFile);
-    this.log.log(
-      `enrichMediaFileFromDisk: enriched media file #${mediaFileId} "${normPath}"`,
-    );
-
-    await this.finalizeImportedFile(saved, absPath, dbFile.media);
-
-    void this.mediaServers.dispatch('library.rescan', {
-      title: dbFile.media.title,
-      path: dbFile.media.path,
-    });
-  }
-
-  /**
-   * Post-probe enrichment shared by every import path (grab→import, disk
-   * import, rescan). Runs ffmpeg `cropdetect` and stashes the result into
-   * `streamInfo.video[0].crop`, then pre-extracts embedded text subtitles
-   * to the player cache (so the first playback doesn't pay the extraction
-   * cost — ExoPlayer on Android blocks prepare until every
-   * SubtitleConfiguration URL has been fetched).
-   *
-   * Caller must have already persisted `file.streamInfo` from a fresh
-   * ffprobe of `absPath`. Best-effort: failures are logged but don't throw.
-   */
-  async finalizeImportedFile(
-    file: MediaFile,
-    absPath: string,
-    media: Media,
-  ): Promise<void> {
-    await this.detectAndStoreCrop(file, absPath);
-    this.rebuildSubtitleCacheForFile(file, absPath, media);
-  }
-
-  /**
-   * Re-runs ffmpeg `cropdetect` for a single file and persists the result
-   * in `streamInfo.video[0].crop`. No-op when the file has no video stream.
-   * Best-effort: ffmpeg failures are logged and the row is left unchanged.
-   * The crop value is cleared when detection returns nothing so a re-encoded
-   * file that lost its letterbox doesn't keep stale crop data.
-   */
-  private async detectAndStoreCrop(
-    file: MediaFile,
-    absPath: string,
-  ): Promise<void> {
-    const streamInfo = file.streamInfo;
-    const v = streamInfo?.video?.[0];
-    if (!v) return;
-    try {
-      const crop = await this.ffprobe.detectCrop(
-        absPath,
-        streamInfo.durationSeconds,
-        v.width,
-        v.height,
-        !!v.hdrFormat,
-      );
-      v.crop = crop ?? undefined;
-      file.streamInfo = streamInfo;
-      await this.mediaFileRepo.save(file);
-    } catch (err) {
-      this.log.warn(
-        `detectAndStoreCrop: failed for "${absPath}"`,
-        err instanceof Error ? err.stack : err,
-      );
-    }
-  }
-
-  /**
-   * Fire-and-forget clear + warmup of the embedded-subtitle cache for a
-   * file. Called after every import so the first playback doesn't pay the
-   * per-track extraction cost (very visible on ExoPlayer / Android TV,
-   * which blocks `prepare` until every SubtitleConfiguration URL has been
-   * fetched).
-   */
-  private rebuildSubtitleCacheForFile(
-    file: MediaFile,
-    absPath: string,
-    media: Media,
-  ): void {
-    void this.subtitleStream
-      .clearMediaFileSubtitleCache(media?.path, file.id)
-      .then(() =>
-        this.subtitleStream.warmupCache(
-          absPath,
-          media?.path,
-          file.id,
-          file.streamInfo?.subtitles,
-          media?.title,
-        ),
-      );
-  }
-
-  /**
-   * User-triggered "Analyse" action — iterates the media's files and runs
-   * each requested granular operation (cropdetect, embedded-subtitle cache
-   * rebuild, forced sprite-sheet regeneration). The full-rescan superset
-   * stays on the dedicated `rescanFiles` path so its SSE wiring lives in
-   * one place.
-   */
-  async analyzeMedia(mediaId: number, opts: AnalyzeMediaDto): Promise<void> {
-    if (!opts.sprites && !opts.crop && !opts.subtitleCache) return;
-
-    const media = await this.mediaRepo.findOne({
-      where: { id: mediaId },
-      relations: ['files', 'files.episode', 'files.episode.season'],
-    });
-    if (!media) throw new NotFoundException(`Media #${mediaId} not found`);
-    if (!media.path) {
-      throw new BadRequestException(
-        `Media #${mediaId} has no root path configured`,
-      );
-    }
-
-    const mediaDir = path.resolve(media.path);
-    const files = media.files ?? [];
-    for (const file of files) {
-      if (!file.relativePath) continue;
-      const absPath = path.join(
-        mediaDir,
-        file.relativePath.replace(/\\/g, '/'),
-      );
-      if (!fs.existsSync(absPath)) {
-        this.log.warn(
-          `analyzeMedia[#${mediaId}]: file off-disk, skipping "${absPath}"`,
-        );
-        continue;
-      }
-      if (opts.crop) {
-        await this.detectAndStoreCrop(file, absPath);
-      }
-      if (opts.subtitleCache) {
-        this.rebuildSubtitleCacheForFile(file, absPath, media);
-      }
-      if (opts.sprites) {
-        const ep = file.episode;
-        const label = buildSpriteLabel(
-          media,
-          ep
-            ? {
-                seasonNumber: ep.season?.seasonNumber,
-                episodeNumber: ep.episodeNumber,
-                title: ep.title,
-              }
-            : null,
-        );
-        // Background, force=true so the existing sprite/meta gets rebuilt.
-        void this.thumbnailService.generateForFile(file, media, label, {
-          force: true,
-        });
-      }
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Rescan files on disk
-  // ---------------------------------------------------------------------------
-
-  private static readonly VIDEO_EXTS = new Set([
-    '.mkv',
-    '.mp4',
-    '.avi',
-    '.mov',
-    '.ts',
-    '.m2ts',
-    '.wmv',
-    '.flv',
-  ]);
-
-  async rescanFiles(
-    mediaId: number,
-    options?: { skipWarmup?: boolean },
-  ): Promise<{
-    added: number;
-    removed: number;
-    updated: number;
-    subtitleRemovedMissing: number;
-    subtitleRemovedDuplicates: number;
-  }> {
-    const media = await this.mediaRepo.findOne({
-      where: { id: mediaId },
-      relations: ['files'],
-    });
-    if (!media) throw new NotFoundException(`Media #${mediaId} not found`);
-    if (!media.path) {
-      throw new BadRequestException(
-        `Media #${mediaId} has no root path configured`,
-      );
-    }
-
-    // Wipe the whole per-media `.cache/` tree (subtitles + any other cached
-    // artefact) — streamInfo is about to be re-read so anything derived
-    // from the old layout is stale.
-    try {
-      await clearMediaCache(media.path);
-    } catch (err) {
-      this.log.warn(
-        `Rescan[media #${mediaId}]: clearMediaCache failed for "${media.path}": ${err instanceof Error ? err.message : err}`,
-      );
-    }
-
-    const mediaDir = path.resolve(media.path);
-    if (!fs.existsSync(mediaDir)) {
-      try {
-        fs.mkdirSync(mediaDir, { recursive: true });
-        this.log.warn(
-          `Rescan: created missing media folder — "${mediaDir}" (media #${mediaId} "${media.title}")`,
-        );
-      } catch (err) {
-        const msg = err instanceof Error ? err.message : String(err);
-        throw new BadRequestException(
-          `Cannot create media folder "${mediaDir}": ${msg}`,
-        );
-      }
-    }
-
-    this.log.log(
-      `Rescan: started — media #${mediaId} "${media.title}" root="${mediaDir}"`,
-    );
-
-    // 1. Collect all video files on disk
-    const rawDiskFiles = this.collectVideoFilesRecursive(mediaDir, 0, mediaId);
-    const diskFiles: string[] = [];
-    const diskRelPaths = new Set<string>();
-    for (const f of rawDiskFiles) {
-      const rel = relativePathUnderMediaRoot(mediaDir, f);
-      if (!rel) {
-        this.log.error(
-          `Rescan[media #${mediaId}]: file is outside resolved media folder — mediaDir="${mediaDir}" file="${f}"`,
-        );
-        continue;
-      }
-      diskRelPaths.add(rel);
-      diskFiles.push(f);
-    }
-    this.log.log(
-      `Rescan: found ${diskFiles.length} file(s) on disk, ${(media.files ?? []).length} in DB`,
-    );
-
-    // 2. Existing DB records
-    const dbFiles = media.files ?? [];
-    if (diskFiles.length === 0 && dbFiles.length > 0) {
-      this.log.warn(
-        `Rescan[media #${mediaId}]: no video file on disk, ${dbFiles.length} file(s) still in DB (orphan rows will be removed if paths do not match)`,
-      );
-    }
-    const dbRelPaths = new Set(
-      dbFiles.map((f) => f.relativePath.replace(/\\/g, '/')),
-    );
-
-    let added = 0;
-    let removed = 0;
-
-    // 3. Remove DB records whose files no longer exist on disk
-    for (const dbFile of dbFiles) {
-      const normPath = dbFile.relativePath?.replace(/\\/g, '/');
-      if (!normPath || !diskRelPaths.has(normPath)) {
-        if (normPath?.includes('..')) {
-          this.log.error(
-            `Rescan[media #${mediaId}]: dropping DB file row with unsafe relativePath (not on disk or invalid): "${normPath}"`,
-          );
-        }
-        const episodeId = dbFile.episodeId;
-        try {
-          await this.mediaFileRepo.remove(dbFile);
-          removed++;
-          this.log.log(
-            `Rescan: removed missing file "${normPath}" for media #${mediaId}`,
-          );
-          if (episodeId != null) {
-            const remaining = await this.mediaFileRepo.count({
-              where: { episode: { id: episodeId } },
-            });
-            if (remaining === 0) {
-              await this.episodeRepo.update(episodeId, { hasFile: false });
-            }
-          }
-        } catch (err) {
-          this.log.error(
-            `Rescan[media #${mediaId}]: failed to remove DB row for missing file "${normPath}"`,
-            err instanceof Error ? err.stack : err,
-          );
-        }
-      }
-    }
-
-    if (removed > 0) {
-      this.log.warn(
-        `Rescan[media #${mediaId}]: removed ${removed} file row(s) from DB (not found on disk)`,
-      );
-    }
-
-    // 4. Refresh metadata for existing DB records from disk
-    let updated = 0;
-    for (const dbFile of dbFiles) {
-      const normPath = dbFile.relativePath?.replace(/\\/g, '/');
-      if (!normPath || !diskRelPaths.has(normPath)) continue;
-      const absPath = path.join(mediaDir, normPath);
-
-      let diskSize: number;
-      try {
-        diskSize = fs.statSync(absPath).size;
-      } catch (err) {
-        this.log.warn(
-          `Rescan[media #${mediaId}]: cannot stat file for refresh (skipped) — path="${absPath}" relativePath="${normPath}"`,
-          err instanceof Error ? err.stack : err,
-        );
-        continue;
-      }
-
-      const filename = path.basename(absPath);
-
-      // Series files: parse filename, either link a missing episodeId OR
-      // patch an existing link with a newly-discovered multi-episode range
-      // (e.g. "S07E11-E12.mkv" already imported as E11 gets endEpisodeNumber
-      // set retroactively).
-      if (media.type === MediaType.SERIES) {
-        const epNums = this.naming.parseEpisodeNumbers(filename);
-        if (epNums && dbFile.episodeId == null) {
-          try {
-            let season = await this.seasonRepo.findOne({
-              where: { media: { id: media.id }, seasonNumber: epNums.season },
-            });
-            if (!season) {
-              season = await this.seasonRepo.save(
-                this.seasonRepo.create({
-                  media,
-                  seasonNumber: epNums.season,
-                  monitored: true,
-                }),
-              );
-              this.log.log(
-                `Rescan: created season ${epNums.season} for media #${mediaId}`,
-              );
-            }
-            let ep = await this.episodeRepo.findOne({
-              where: {
-                season: { id: season.id },
-                episodeNumber: epNums.episode,
-              },
-            });
-            if (!ep) {
-              ep = await this.episodeRepo.save(
-                this.episodeRepo.create({
-                  season,
-                  episodeNumber: epNums.episode,
-                  endEpisodeNumber: epNums.episodeEnd ?? null,
-                  monitored: true,
-                }),
-              );
-              this.log.log(
-                `Rescan: created episode S${String(epNums.season).padStart(2, '0')}E${String(epNums.episode).padStart(2, '0')} for media #${mediaId}`,
-              );
-            } else if (
-              epNums.episodeEnd != null &&
-              ep.endEpisodeNumber !== epNums.episodeEnd
-            ) {
-              ep.endEpisodeNumber = epNums.episodeEnd;
-              await this.episodeRepo.save(ep);
-            }
-            dbFile.episode = ep;
-            try {
-              await this.mediaFileRepo.save(dbFile);
-              await this.episodeRepo.update(ep.id, { hasFile: true });
-              updated++;
-              this.log.log(
-                `Rescan: linked "${normPath}" to S${String(epNums.season).padStart(2, '0')}E${String(epNums.episode).padStart(2, '0')} for media #${mediaId}`,
-              );
-            } catch (err) {
-              this.log.error(
-                `Rescan[media #${mediaId}]: failed to link file "${normPath}" to episode`,
-                err instanceof Error ? err.stack : err,
-              );
-            }
-          } catch (err) {
-            this.log.error(
-              `Rescan[media #${mediaId}]: failed to create season/episode for refresh "${normPath}"`,
-              err instanceof Error ? err.stack : err,
-            );
-          }
-        } else if (epNums?.episodeEnd != null && dbFile.episodeId != null) {
-          // Already-linked file: retroactively apply the range on its Episode.
-          const linked = await this.episodeRepo.findOne({
-            where: { id: dbFile.episodeId },
-          });
-          if (linked && linked.endEpisodeNumber !== epNums.episodeEnd) {
-            linked.endEpisodeNumber = epNums.episodeEnd;
-            await this.episodeRepo.save(linked);
-            this.log.log(
-              `Rescan: set endEpisodeNumber=${epNums.episodeEnd} on S${String(epNums.season).padStart(2, '0')}E${String(epNums.episode).padStart(2, '0')} for media #${mediaId}`,
-            );
-          }
-        }
-      }
-
-      // Always re-probe streamInfo (fast, ~1s) to pick up schema changes.
-      // Skip detectCrop (~5-10s parallel) when the file size hasn't
-      // changed — crop is a property of the file, so unchanged bytes
-      // mean unchanged crop. Backfilling files that lack crop metadata
-      // would otherwise force every legacy rescan to do the slow probe
-      // again. Operators can force re-detection by deleting the dbFile
-      // row or temporarily breaking the size match.
-      const sizeUnchanged = dbFile.size === diskSize;
-      dbFile.size = diskSize;
-      let streamInfo: Awaited<
-        ReturnType<FfprobeService['detectMediaFileInfo']>
-      >;
-      try {
-        streamInfo = await this.ffprobe.detectMediaFileInfo(absPath);
-        if (streamInfo?.video?.[0] && !sizeUnchanged) {
-          try {
-            const v = streamInfo.video[0];
-            const crop = await this.ffprobe.detectCrop(
-              absPath,
-              streamInfo.durationSeconds,
-              v.width,
-              v.height,
-              !!v.hdrFormat,
-            );
-            if (crop) v.crop = crop;
-          } catch (err) {
-            this.log.warn(
-              `Rescan[media #${mediaId}]: detectCrop failed on refresh "${normPath}" abs="${absPath}" (metadata otherwise kept)`,
-              err instanceof Error ? err.stack : err,
-            );
-          }
-        } else if (
-          sizeUnchanged &&
-          (dbFile.streamInfo as any)?.video?.[0]?.crop
-        ) {
-          // Preserve existing crop from previous scan
-          streamInfo.video[0].crop = (dbFile.streamInfo as any).video[0].crop;
-        }
-        dbFile.streamInfo = streamInfo;
-      } catch (err) {
-        this.log.error(
-          `Rescan[media #${mediaId}]: ffprobe failed on refresh "${normPath}" abs="${absPath}"`,
-          err instanceof Error ? err.stack : err,
-        );
-        continue;
-      }
-      const qualityName = this.resolveQuality(
-        filename,
-        streamInfo?.video?.[0]?.height,
-        streamInfo?.video?.[0]?.width,
-      );
-      dbFile.quality = qualityName;
-      try {
-        await this.mediaFileRepo.save(dbFile);
-        updated++;
-        this.log.log(
-          `Rescan: refreshed "${normPath}" for media #${mediaId} (size: ${diskSize}, quality: ${qualityName}${sizeUnchanged ? ', skipped crop' : ''})`,
-        );
-        if (!options?.skipWarmup) {
-          void this.subtitleStream.warmupCache(
-            absPath,
-            media.path,
-            dbFile.id,
-            streamInfo?.subtitles,
-            media.title,
-          );
-        }
-      } catch (err) {
-        this.log.error(
-          `Rescan[media #${mediaId}]: failed to save refreshed metadata for "${normPath}"`,
-          err instanceof Error ? err.stack : err,
-        );
-      }
-    }
-
-    // 5. Add new files found on disk but not in DB
-    for (const absPath of diskFiles) {
-      const relativePath = relativePathUnderMediaRoot(mediaDir, absPath);
-      if (!relativePath) {
-        this.log.error(
-          `Rescan[media #${mediaId}]: internal inconsistency — file was listed but not under mediaDir — mediaDir="${mediaDir}" file="${absPath}"`,
-        );
-        continue;
-      }
-      if (dbRelPaths.has(relativePath)) continue;
-
-      let size = 0;
-      try {
-        size = fs.statSync(absPath).size;
-      } catch (err) {
-        this.log.error(
-          `Rescan[media #${mediaId}]: cannot stat new file — path="${absPath}"`,
-          err instanceof Error ? err.stack : err,
-        );
-        continue;
-      }
-
-      const filename = path.basename(absPath);
-
-      // Try to match episode for series — create season/episode on the fly if missing
-      let episodeId: number | undefined;
-      if (media.type === MediaType.SERIES) {
-        const epNums = this.naming.parseEpisodeNumbers(filename);
-        if (epNums) {
-          try {
-            let season = await this.seasonRepo.findOne({
-              where: { media: { id: media.id }, seasonNumber: epNums.season },
-            });
-            if (!season) {
-              season = await this.seasonRepo.save(
-                this.seasonRepo.create({
-                  media,
-                  seasonNumber: epNums.season,
-                  monitored: true,
-                }),
-              );
-              this.log.log(
-                `Rescan: created season ${epNums.season} for media #${mediaId}`,
-              );
-            }
-            let ep = await this.episodeRepo.findOne({
-              where: {
-                season: { id: season.id },
-                episodeNumber: epNums.episode,
-              },
-            });
-            if (!ep) {
-              ep = await this.episodeRepo.save(
-                this.episodeRepo.create({
-                  season,
-                  episodeNumber: epNums.episode,
-                  endEpisodeNumber: epNums.episodeEnd ?? null,
-                  monitored: true,
-                }),
-              );
-              this.log.log(
-                `Rescan: created episode S${String(epNums.season).padStart(2, '0')}E${String(epNums.episode).padStart(2, '0')} for media #${mediaId}`,
-              );
-            } else if (
-              epNums.episodeEnd != null &&
-              ep.endEpisodeNumber !== epNums.episodeEnd
-            ) {
-              ep.endEpisodeNumber = epNums.episodeEnd;
-              await this.episodeRepo.save(ep);
-            }
-            episodeId = ep.id;
-          } catch (err) {
-            this.log.error(
-              `Rescan[media #${mediaId}]: failed to create season/episode for new file "${filename}" — importing file without episode link`,
-              err instanceof Error ? err.stack : err,
-            );
-          }
-        } else {
-          this.log.warn(
-            `Rescan[media #${mediaId}]: series file name has no SxxEyy pattern — "${filename}" (skipped, alien file in series folder)`,
-          );
-          continue;
-        }
-      }
-
-      let streamInfo: Awaited<
-        ReturnType<FfprobeService['detectMediaFileInfo']>
-      >;
-      try {
-        streamInfo = await this.ffprobe.detectMediaFileInfo(absPath);
-        if (streamInfo?.video?.[0]) {
-          try {
-            const v = streamInfo.video[0];
-            const crop = await this.ffprobe.detectCrop(
-              absPath,
-              streamInfo.durationSeconds,
-              v.width,
-              v.height,
-              !!v.hdrFormat,
-            );
-            if (crop) v.crop = crop;
-          } catch (err) {
-            this.log.warn(
-              `Rescan[media #${mediaId}]: detectCrop failed for "${relativePath}" abs="${absPath}" (file still imported)`,
-              err instanceof Error ? err.stack : err,
-            );
-          }
-        }
-      } catch (err) {
-        this.log.error(
-          `Rescan[media #${mediaId}]: ffprobe failed for new file "${relativePath}" abs="${absPath}"`,
-          err instanceof Error ? err.stack : err,
-        );
-        continue;
-      }
-      const qualityName = this.resolveQuality(
-        filename,
-        streamInfo.video?.[0]?.height,
-        streamInfo.video?.[0]?.width,
-      );
-      try {
-        const savedFile = await this.mediaFileRepo.save(
-          this.mediaFileRepo.create({
-            media,
-            episode: episodeId != null ? ({ id: episodeId } as Episode) : null,
-            relativePath,
-            size,
-            quality: qualityName,
-            streamInfo,
-          }),
-        );
-        added++;
-        this.log.log(
-          `Rescan: added new file "${relativePath}" for media #${mediaId}`,
-        );
-        if (episodeId != null) {
-          await this.episodeRepo.update(episodeId, { hasFile: true });
-        }
-        if (!options?.skipWarmup) {
-          void this.subtitleStream.warmupCache(
-            absPath,
-            media.path,
-            savedFile.id,
-            streamInfo.subtitles,
-            media.title,
-          );
-        }
-      } catch (err) {
-        this.log.error(
-          `Rescan[media #${mediaId}]: failed to save new file row "${relativePath}" abs="${absPath}"`,
-          err instanceof Error ? err.stack : err,
-        );
-      }
-    }
-
-    let subtitleRemovedMissing = 0;
-    let subtitleRemovedDuplicates = 0;
-    try {
-      const sub =
-        await this.subtitles.reconcileSubtitleFilesAfterRescan(mediaId);
-      subtitleRemovedMissing = sub.removedMissing;
-      subtitleRemovedDuplicates = sub.removedDuplicates;
-      if (subtitleRemovedMissing || subtitleRemovedDuplicates) {
-        this.log.log(
-          `Rescan: subtitles reconciled — media #${mediaId} removedMissing=${subtitleRemovedMissing} removedDuplicates=${subtitleRemovedDuplicates}`,
-        );
-      }
-      // Discover external subtitle files on disk (filename-based)
-      const discovered =
-        await this.subtitles.discoverExternalSubtitles(mediaId);
-      if (discovered) {
-        this.log.log(
-          `Rescan: discovered ${discovered} external subtitle(s) on disk for media #${mediaId}`,
-        );
-      }
-      // Always re-detect embedded subtitle streams on rescan: detectAndStore
-      // wipes the file's existing embedded rows and recreates them from
-      // ffprobe, so deleted tracks disappear and new ones get added even
-      // when the file size hasn't changed (e.g. a remux that preserves size).
-      const allFiles = await this.mediaFileRepo.find({
-        where: { media: { id: mediaId } },
-      });
-      for (const file of allFiles) {
-        try {
-          await this.embeddedSubtitle.detectAndStore(
-            mediaId,
-            file.id,
-            file.episodeId ?? undefined,
-          );
-        } catch (err) {
-          this.log.warn(
-            `Rescan: embedded subtitle detect failed for file #${file.id} — ${err instanceof Error ? err.message : err}`,
-          );
-        }
-      }
-    } catch (err) {
-      this.log.warn(
-        `Rescan: subtitle reconcile/discover failed for media #${mediaId} — ${err instanceof Error ? err.message : err}`,
-      );
-    }
-
-    this.log.log(
-      `Rescan: finished — media #${mediaId} "${media.title}" added=${added} removed=${removed} updated=${updated}`,
-    );
-    if (added === 0 && removed === 0 && updated === 0) {
-      this.log.warn(
-        `Rescan[media #${mediaId}]: no changes (added=0 removed=0 updated=0)`,
-      );
-    }
-
-    if (added || removed || updated) {
-      void this.mediaServers.dispatch('library.rescan', {
-        title: media.title,
-        path: media.path,
-      });
-    }
-
-    return {
-      added,
-      removed,
-      updated,
-      subtitleRemovedMissing,
-      subtitleRemovedDuplicates,
-    };
-  }
-
-  private collectVideoFilesRecursive(
-    dir: string,
-    depth: number,
-    mediaId: number,
-  ): string[] {
-    if (depth > 3) {
-      this.log.warn(
-        `Rescan[media #${mediaId}]: skipping subfolder (max depth 3) — "${dir}"`,
-      );
-      return [];
-    }
-    const files: string[] = [];
-    let entries: fs.Dirent[];
-    try {
-      entries = fs.readdirSync(dir, { withFileTypes: true });
-    } catch (err) {
-      this.log.error(
-        `Rescan[media #${mediaId}]: cannot read directory (permissions, missing path, or I/O) — "${dir}"`,
-        err instanceof Error ? err.stack : err,
-      );
-      return [];
-    }
-    for (const entry of entries) {
-      const fullPath = path.join(dir, entry.name);
-      if (entry.isDirectory()) {
-        files.push(
-          ...this.collectVideoFilesRecursive(fullPath, depth + 1, mediaId),
-        );
-      } else if (
-        MediaService.VIDEO_EXTS.has(path.extname(entry.name).toLowerCase())
-      ) {
-        files.push(fullPath);
-      }
-    }
-    return files;
-  }
-
-  /**
-   * Determine quality from ffprobe resolution (source of truth) + filename source tag.
-   */
-  private resolveQuality(
-    filename: string,
-    actualHeight?: number,
-    actualWidth?: number,
-  ): string {
-    // Resolution bucket from frame dimensions: first ceiling that fits BOTH
-    // width AND height wins. Ceilings on both axes absorb anamorphic / scope
-    // crops (e.g. a 1080p source encoded as 1796x1076 or 1920x800) which a
-    // width-only threshold mis-buckets as 720p.
-    const w = actualWidth ?? 0;
-    const h = actualHeight ?? 0;
-    let resolution: number;
-    if (!w && !h) resolution = 480;
-    else if (w <= 720 && h <= 576) resolution = 480;
-    else if (w <= 1280 && h <= 962) resolution = 720;
-    else if (w <= 1920 && h <= 1440) resolution = 1080;
-    else resolution = 2160;
-
-    // Determine source from filename (bluray, web, remux, etc.)
-    const t = filename.replace(/\./g, ' ').toLowerCase();
-    let source = 'hdtv';
-    if (/\bremux\b/.test(t)) source = 'remux';
-    else if (/\b(bluray|blu-?ray|bdrip|brrip)\b/.test(t)) source = 'bluray';
-    else if (/\bweb-?dl\b/.test(t)) source = 'web';
-    else if (/\bweb-?rip\b/.test(t)) source = 'web';
-    else if (/\b(dvd|dvdrip)\b/.test(t)) source = 'dvd';
-
-    const match = APP_QUALITIES.find(
-      (q) => q.resolution === resolution && q.source === source,
-    );
-    if (match) return match.name;
-
-    // Fallback: any quality with correct resolution
-    const fallback = APP_QUALITIES.find((q) => q.resolution === resolution);
-    return fallback?.name ?? `HDTV-${resolution}p`;
   }
 }
